% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ActivityEvents-class}
\alias{ActivityEvents-class}
\alias{.ActivityEvents}
\title{Activity}
\description{
An S4 class to store the result of an \link{activity} plot.
}
\section{Slots}{

\describe{
\item{\code{hash}}{A \code{\link{character}} string giving the 32-byte MD5 hash of the
original data file.}
}}

\note{
This class inherits from \code{\link[aion:TimeSeries-class]{aion::TimeSeries}}.
}
\section{Coerce}{

In the code snippets below, \code{x} is an \code{ActivityEvents} object.
\describe{
\item{\code{as.data.frame(x)}}{Coerces to a \code{\link{data.frame}}.}
}
}

\seealso{
Other classes:
\code{\link{AgeDepthModel-class}},
\code{\link{CumulativeEvents-class}},
\code{\link{DurationsMCMC-class}},
\code{\link{EventsMCMC-class}},
\code{\link{MCMC-class}},
\code{\link{OccurrenceEvents-class}},
\code{\link{PhasesMCMC-class}},
\code{\link{TimeRange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
