% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AgeDepthModel-class}
\alias{AgeDepthModel-class}
\alias{.AgeDepthModel}
\title{Age-Depth Model}
\description{
An S4 class to represents an age-depth model.
}
\section{Slots}{

\describe{
\item{\code{depth}}{A \code{\link{numeric}} vector giving the depth of the samples.}

\item{\code{model}}{A \code{\link{list}} of local polynomial regressions
(see \code{\link[stats:loess]{stats::loess()}}).}

\item{\code{calendar}}{A \code{\link{character}} string specifying the chronological scale
(either "\code{CE}", "\code{BP}" or "\code{b2k}").}

\item{\code{hash}}{A \code{\link{character}} string giving the 32-byte MD5 hash of the
original data file.}
}}

\seealso{
Other classes: 
\code{\link{ActivityEvents-class}},
\code{\link{CumulativeEvents-class}},
\code{\link{MCMC}},
\code{\link{OccurrenceEvents-class}},
\code{\link{PhasesMCMC-class}},
\code{\link{ProxyRecord-class}},
\code{\link{RECE-class}},
\code{\link{RateOfChange-class}},
\code{\link{TimeRange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
