% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSet.R
\name{allen.basic.relation.set}
\alias{allen.basic.relation.set}
\title{The basic Allen relation set}
\usage{
allen.basic.relation.set()
}
\value{
A vector of thirteen one-letter codes
}
\description{
A vector of one-letter codes for the thirteen basic Allen relations.
The codes were proposed by Thomas Alspaugh.
}
\author{
Thomas S. Dye
}
