% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenVector.R
\name{allen.calculate.relations.2}
\alias{allen.calculate.relations.2}
\title{Calculate the Allen relation of two intervals}
\usage{
allen.calculate.relations.2(result.vector, mcmc.chain.list)
}
\arguments{
\item{result.vector}{A result vector}

\item{mcmc.chain.list}{A list with four MCMC chains representing the
beginning and ending boundaries of two intervals.}
}
\value{
An updated result vector
}
\description{
Given a result vector and four MCMC chains, calculate the Allen relation
and update the result vector with the observed relation.
}
\author{
Thomas S. Dye
}
