% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSupport.R
\name{allen.check.names}
\alias{allen.check.names}
\title{Extract phase names from MCMC csv file column headers.}
\usage{
allen.check.names(names)
}
\arguments{
\item{names}{A vector of four MCMC csv file column headers.}
}
\value{
A list of two phase names.
}
\description{
Extract two phase names from four MCMC csv file column headers, checking that
each pair refers to a single phase.  Stops with an error if the names in a
pair don't refer to the same phase.
}
\author{
Thomas S. Dye
}
