% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSet.R
\name{allen.complement.set}
\alias{allen.complement.set}
\title{Allen basic relation set complement}
\usage{
allen.complement.set(allen.set)
}
\arguments{
\item{allen.set}{An Allen set or result vector}
}
\value{
An Allen set, a vector of one-letter codes.
}
\description{
Returns the complement of the set passed as an argument.
}
\author{
Thomas S. Dye
}
