% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenComposition.R
\name{allen.composition}
\alias{allen.composition}
\title{Composition of two Allen relation sets}
\usage{
allen.composition(allen.set.1, allen.set.2)
}
\arguments{
\item{allen.set.1}{the first Allen relation set}

\item{allen.set.2}{the second Allen relation set}
}
\description{
Composition of two Allen relation sets
}
\author{
Thomas S. Dye
}
