% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSet.R
\name{allen.concurrent.relation.set}
\alias{allen.concurrent.relation.set}
\title{Allen concurrent relation set}
\usage{
allen.concurrent.relation.set()
}
\value{
A vector of nine one-letter codes.
}
\description{
A vector of nine one-letter codes for the Allen concurrent relations.
The codes were proposed by Thomas Alspaugh.
}
