% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSet.R
\name{allen.converse.set}
\alias{allen.converse.set}
\title{Allen relation set converse.}
\usage{
allen.converse.set(allen.set)
}
\arguments{
\item{allen.set}{An Allen relation set or result vector}
}
\value{
An Allen relation set
}
\description{
Calculates the converse of an Allen relation set, taking care to
convert a result vector to a relation set.
}
\author{
Thomas S. Dye
}
