% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenVector.R
\name{allen.create.concurrent.vector}
\alias{allen.create.concurrent.vector}
\title{Create a result vector identifying concurrent relations}
\usage{
allen.create.concurrent.vector()
}
\value{
A result vector
}
\description{
Create a result vector where concurrent relations are set to 1
and non-concurrent relations are set to 0.
}
\author{
Thomas S. Dye
}
