% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenVector.R
\name{allen.create.result.vector}
\alias{allen.create.result.vector}
\title{Create a named result vector}
\usage{
allen.create.result.vector(initial.value = 0)
}
\arguments{
\item{initial.value}{A value used to initialize the vector. typically 0
(default) or 1.}
}
\value{
An initialized result vector.
}
\description{
Create a named result vector initialized to zero by default or to some
other value.
}
\author{
Thomas S. Dye
}
