% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSupport.R
\name{allen.order.string}
\alias{allen.order.string}
\title{Order Allen relations in a string representing an Allen set}
\usage{
allen.order.string(s)
}
\arguments{
\item{s}{A string representing an Allen set}
}
\value{
A string ordered conventionally
}
\description{
Order Allen relations in a string representing an Allen set
}
\author{
Thomas S. Dye
}
