% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenIo.R
\name{allen.read.mcmc.chains}
\alias{allen.read.mcmc.chains}
\title{Read MCMC chains from Bayesian calibration software output}
\usage{
allen.read.mcmc.chains(
  mcmc.file,
  positions,
  mcmc.file.2 = NULL,
  positions.2 = NULL,
  names = c("Interval 1", "Interval 2"),
  app = "bcal",
  decimal = ".",
  separator = ",",
  quiet = "partial"
)
}
\arguments{
\item{mcmc.file}{path to the csv file of MCMC output}

\item{positions}{a vector of 2 or 4 column indices of \code{mcmc.file}}

\item{mcmc.file.2}{optional path to another csv file of MCMC output}

\item{positions.2}{a vector of 2 column indices of \code{mcmc.file.2}}

\item{names}{a vector of two strings naming the intervals specified by
the \code{positions} parameters(s)}

\item{app}{one of 'bcal', 'oxcal', or 'chronomodel'}

\item{decimal}{optional decimal character for Chronomodel output}

\item{separator}{optional separator character for Chronomodel output}

\item{quiet}{One of 'no' to allow messages and warnings,
'partial' (default) to suppress messages and allow warnings, or 'yes'
to suppress messages and warnings.}
}
\value{
a list with two components: \code{result} and \code{names}, where \code{result}
contains the MCMC chains and \code{names} contains the names parameter.
}
\description{
Reads chains for two intervals from one or two csv files of MCMC
output from a Bayesian calibration application.
The function checks for missing or incorrect parameters and whether
or not files exist. It exits with an error message if there are problems.
}
\author{
Thomas S. Dye
}
