% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSupport.R
\name{allen.relation}
\alias{allen.relation}
\title{Allen relation of two definite intervals}
\usage{
allen.relation(start.1, end.1, start.2, end.2)
}
\arguments{
\item{start.1}{The start date of the first interval}

\item{end.1}{The end date of the first interval}

\item{start.2}{The start date of the second interval}

\item{end.2}{The end date of the second interval}
}
\value{
A one-letter code indicating the Allen relation
}
\description{
Calculates the Allen relation of two definite intervals and
reports the one-letter code for the interval proposed by Thomas Alspaugh.
Stops with an error if the end of an interval is earlier than its start.
}
\author{
Thomas S. Dye
}
