% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenIo.R
\name{allen.relations.concur}
\alias{allen.relations.concur}
\title{Calculate the concurrence set of an Allen relation sets}
\usage{
allen.relations.concur(allen.set)
}
\arguments{
\item{allen.set}{Allen set}
}
\value{
A string that describes the concurrence relation
}
\description{
Return a string that describes the concurrence set of an Allen relation set,
or indicates there is no concurrence set.  Useful for output
to the console.
}
\author{
Thomas S. Dye
}
