% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSet.R
\name{allen.relations.intersection}
\alias{allen.relations.intersection}
\title{Intersection of Allen relation sets}
\usage{
allen.relations.intersection(allen.set.1, allen.set.2)
}
\arguments{
\item{allen.set.1}{An Allen set}

\item{allen.set.2}{An Allen set}
}
\value{
An Allen relation set
}
\description{
Calculates the intersection of two Allen relation sets.  Attempts to
handle sets represented by result vectors, as well as sets represented
by vectors of one-letter codes.
}
\author{
Thomas S. Dye
}
