% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenIo.R
\name{allen.relations.relationships}
\alias{allen.relations.relationships}
\title{Relative strength of Allen relation sets}
\usage{
allen.relations.relationships(allen.set.1, allen.set.2)
}
\arguments{
\item{allen.set.1}{an Allen set}

\item{allen.set.2}{an Allen set}
}
\value{
A string that describes the relative strength of two Allen relation sets.
}
\description{
Construct a string that describes the relative strength of two Allen relation
sets.  Useful for output to the console.
}
\author{
Thomas S. Dye
}
