% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenIo.R
\name{allen.relations.set}
\alias{allen.relations.set}
\title{Express the Allen relation set in mathematical notation}
\usage{
allen.relations.set(allen.set)
}
\arguments{
\item{allen.set}{an Allen relation set}
}
\description{
Return a string expressing an Allen relation set in mathematical
set notation.  Useful for printing to the console.
}
\author{
Thomas S. Dye
}
