% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSet.R
\name{allen.relations.union}
\alias{allen.relations.union}
\title{Union of two Allen relation sets.}
\usage{
allen.relations.union(allen.set.1, allen.set.2)
}
\arguments{
\item{allen.set.1}{The first Allen relation set or result vector}

\item{allen.set.2}{The second Allen relation set or result vector}
}
\value{
An Allen relation set
}
\description{
Returns the union of two Allen relation sets, taking care to handle
empty sets and the sets represented by result vectors.
}
\author{
Thomas S. Dye
}
