% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSupport.R
\name{allen.set.to.vector}
\alias{allen.set.to.vector}
\title{Convert an Allen relation set to a named vector}
\usage{
allen.set.to.vector(s)
}
\arguments{
\item{s}{An Allen relation set, a vector of single letter codes.}
}
\value{
A named result vector.
}
\description{
Set elements that are not Allen relation codes are silently ignored.
}
\author{
Thomas S. Dye
}
