% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSet.R
\name{allen.set.vector.from.result}
\alias{allen.set.vector.from.result}
\title{The Allen relation set represented by the non-zero elements of a result vector}
\usage{
allen.set.vector.from.result(result.vector)
}
\arguments{
\item{result.vector}{A result vector}
}
\value{
An Allen set
}
\description{
Given a result vector, returns a vector of the one-letter codes corresponding to
the non-zero elements of the result vector.
}
\author{
Thomas S. Dye
}
