% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSet.R
\name{allen.set.vector}
\alias{allen.set.vector}
\title{Coerce a string to an Allen relation set.}
\usage{
allen.set.vector(allen.set.string)
}
\arguments{
\item{allen.set.string}{A string whose elements are expected to be one-letter
codes for Allen relations following the convention proposed by Thomas
Alspaugh.}
}
\description{
Coerces a string to a vector representing an Allen relation set.
Does not check if the string represents an Allen set correctly.
}
\author{
Thomas S. Dye
}
