% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSet.R
\name{allen.six.value.set}
\alias{allen.six.value.set}
\title{Allen relation set for intervals with distinct endpoints.}
\usage{
allen.six.value.set()
}
\value{
An Allen relation set
}
\description{
Return the six value Allen relation set for intervals with distinct
endpoints.
}
\author{
Thomas S. Dye
}
