% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSet.R
\name{allen.string.from.result}
\alias{allen.string.from.result}
\title{String representation of an Allen set from a result vector}
\usage{
allen.string.from.result(result.vector)
}
\arguments{
\item{result.vector}{A result vector}
}
\value{
A string representing an Allen set
}
\description{
Given a result vector, return a string corresponding to the non-zero
elements of the result vector
}
\author{
Thomas S. Dye
}
