% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenVector.R
\name{allen.update.result}
\alias{allen.update.result}
\title{Update a result vector}
\usage{
allen.update.result(result.vector, relation)
}
\arguments{
\item{result.vector}{The result vector to update}

\item{relation}{The relation to increment}
}
\value{
The updated result vector
}
\description{
Increment the element of the result vector corresponding to the given
relation.
}
\author{
Thomas S. Dye
}
