% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allen-mcmc.R
\name{allen_analyze_relations}
\alias{allen_analyze_relations}
\title{Data for an Analytic Graphic}
\usage{
allen_analyze_relations(x, y)
}
\arguments{
\item{x, y}{A \code{\link{character}} string denoting an Allen relation.}
}
\value{
A \code{data.frame} to be passed to \code{\link[=allen_plot]{allen_plot()}}.
}
\description{
Calculates the Allen composition of two relations.
}
\seealso{
Other Allen's intervals:
\code{\link[=allen_analyze]{allen_analyze()}},
\code{\link[=allen_complement]{allen_complement()}},
\code{\link[=allen_composition]{allen_composition()}},
\code{\link[=allen_converse]{allen_converse()}},
\code{\link[=allen_illustrate]{allen_illustrate()}},
\code{\link[=allen_illustrate_relations]{allen_illustrate_relations()}},
\code{\link[=allen_intersect]{allen_intersect()}},
\code{\link[=allen_joint_concurrency]{allen_joint_concurrency()}},
\code{\link[=allen_observe]{allen_observe()}},
\code{\link[=allen_observe_frequency]{allen_observe_frequency()}},
\code{\link[=allen_plot]{allen_plot()}},
\code{\link[=allen_relation]{allen_relation()}},
\code{\link[=allen_relation_code]{allen_relation_code()}},
\code{\link[=allen_union]{allen_union()}}
}
\author{
T. S. Dye
}
\concept{Allen's intervals}
\keyword{internal}
