% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_basic_relation_set}
\alias{allen_basic_relation_set}
\title{The basic Allen relation set}
\usage{
allen_basic_relation_set()
}
\value{
A vector of thirteen one-letter codes
}
\description{
A vector of one-letter codes for the thirteen basic Allen relations.
The codes were proposed by Thomas Alspaugh.
}
\author{
Thomas S. Dye
}
