% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSet.R
\name{allen_basic_relation_strings}
\alias{allen_basic_relation_strings}
\title{Allen basic relation set as strings}
\usage{
allen_basic_relation_strings()
}
\value{
A vector of thirteen strings
}
\description{
String descriptors of the Allen basic relations.
}
\author{
Thomas S. Dye
}
