% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSupport.R
\name{allen_code_to_string}
\alias{allen_code_to_string}
\title{Relationship string from code}
\usage{
allen_code_to_string(code)
}
\arguments{
\item{code}{A one-letter code}
}
\value{
A string
}
\description{
Return a string that describes the relationship indicated by a one-letter code.
Stops with an error if the code is not recognized.
}
\author{
Thomas S. Dye
}
