% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_concurrent_relations}
\alias{allen_concurrent_relations}
\title{Allen concurrent relation set}
\usage{
allen_concurrent_relations()
}
\value{
A vector of nine one-letter codes.
}
\description{
A vector of nine one-letter codes for the Allen concurrent relations.
The codes were proposed by Thomas Alspaugh.
}
