% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allen-mcmc.R
\name{allen_count}
\alias{allen_count}
\title{Count Allen Relations}
\usage{
allen_count(x, count = TRUE)
}
\value{
A \code{\link{data.frame}} with two columns: "\code{code}" and "\code{result}".
}
\description{
Count Allen Relations
}
\author{
N. Frerebeau
}
\keyword{internal}
