% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_create_concurrent_vector}
\alias{allen_create_concurrent_vector}
\title{Create a result vector identifying concurrent relations}
\usage{
allen_create_concurrent_vector()
}
\value{
A result vector
}
\description{
Create a result vector where concurrent relations are set to 1
and non-concurrent relations are set to 0.
}
\author{
Thomas S. Dye
}
