% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_create_distinct_endpoint_vector}
\alias{allen_create_distinct_endpoint_vector}
\title{Create a result vector for relations with distinct endpoints}
\usage{
allen_create_distinct_endpoint_vector()
}
\value{
A named vector with distinct endpoint relations set to 1
and all others set to 0.
}
\description{
The six relations with distinct endpoints are commonly observed when
comparing indefinite intervals, such as those returned by a Bayesian
calibration
}
\author{
Thomas S. Dye
}
