% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_create_result_vector}
\alias{allen_create_result_vector}
\title{Create a named result vector}
\usage{
allen_create_result_vector(initial_value = 0)
}
\arguments{
\item{initial_value}{A value used to initialize the vector. typically 0
(default) or 1.}
}
\value{
An initialized result vector.
}
\description{
Create a named result vector initialized to zero by default or to some
other value.
}
\author{
Thomas S. Dye
}
