% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_ensure_set_vector}
\alias{allen_ensure_set_vector}
\title{Ensure an Allen set is represented as a vector of single character strings}
\usage{
allen_ensure_set_vector(obj)
}
\arguments{
\item{obj}{An Allen set represented as a string, a set vector, or a
result vector.}
}
\value{
An Allen set represented as a set vector.
}
\description{
Expects a string, set vector, or result vector and will stop with an
error if something else is encountered.
}
\author{
Thomas S. Dye
}
