% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_is_result_vector}
\alias{allen_is_result_vector}
\title{Test whether an object is a result vector}
\usage{
allen_is_result_vector(obj)
}
\arguments{
\item{obj}{An object to test}
}
\value{
Boolean, TRUE if obj is a result vector, FALSE otherwise.
}
\description{
Checks for vector, names, and class
}
