% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_is_set_string}
\alias{allen_is_set_string}
\title{Test if an object is a set string}
\usage{
allen_is_set_string(obj)
}
\arguments{
\item{obj}{An object to test}
}
\value{
Boolean, TRUE if obj is a set string, FALSE otherwise.
}
\description{
Checks for mode 'character', length of 1, and nchar <= 13
}
\author{
Thomas S. Dye
}
