% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSupport.R
\name{allen_lattice_y}
\alias{allen_lattice_y}
\title{Nokel lattice y coordinates}
\usage{
allen_lattice_y()
}
\value{
A vector of integers
}
\description{
A vector of arbitrary coordinates for lattice node placement
}
\author{
Thomas S. Dye
}
