% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allen-mcmc.R
\name{allen_plot}
\alias{allen_plot}
\title{Make a Single Plot of a Nökel Lattice.}
\usage{
allen_plot(x, main = NULL, sub = NULL, ann = graphics::par("ann"), ...)
}
\arguments{
\item{x}{A \code{\link{data.frame}} with plot information, such as the one
produced by \code{\link[=allen_illustrate_relations]{allen_illustrate_relations()}}.}

\item{main}{A \code{\link{character}} string giving a main title for the plot.}

\item{sub}{A \code{\link{character}} string giving a subtitle for the plot.}

\item{ann}{A \code{\link{logical}} scalar: should the default annotation appear on the
plot?}

\item{...}{Currently not used.}
}
\value{
\code{allen_plot()} is called it for its side-effects: it results in a graphic
being displayed (invisibly returns \code{x}).
}
\description{
Plots a Nökel lattice.
}
\seealso{
Other Allen's intervals:
\code{\link[=allen_analyze]{allen_analyze()}},
\code{\link[=allen_analyze_relations]{allen_analyze_relations()}},
\code{\link[=allen_complement]{allen_complement()}},
\code{\link[=allen_composition]{allen_composition()}},
\code{\link[=allen_converse]{allen_converse()}},
\code{\link[=allen_illustrate]{allen_illustrate()}},
\code{\link[=allen_illustrate_relations]{allen_illustrate_relations()}},
\code{\link[=allen_intersect]{allen_intersect()}},
\code{\link[=allen_joint_concurrency]{allen_joint_concurrency()}},
\code{\link[=allen_observe]{allen_observe()}},
\code{\link[=allen_observe_frequency]{allen_observe_frequency()}},
\code{\link[=allen_relation]{allen_relation()}},
\code{\link[=allen_relation_code]{allen_relation_code()}},
\code{\link[=allen_union]{allen_union()}}
}
\author{
T. S. Dye, N. Frerebeau
}
\concept{Allen's intervals}
\keyword{internal}
