% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenIo.R
\name{allen_relation_summary}
\alias{allen_relation_summary}
\title{Summarize the relation of two phases}
\usage{
allen_relation_summary(mcmc, phases, app = "bcal", quiet = "partial")
}
\arguments{
\item{mcmc}{path to a csv file with MCMC output}

\item{phases}{a vector with four column indices representing the start
and end chains of two phases}

\item{app}{one of 'bcal', 'oxcal', or 'chronomodel' to specify which
Bayesian calibration application produced the MCMC output}

\item{quiet}{One of 'no' to allow messages and warnings,
'partial' (default) to suppress messages and allow warnings, or 'yes'
to suppress messages and warnings.}
}
\value{
a list with the following components:
\describe{
\item{result}{The Allen relation as a string.}
\item{relation_set}{The Allen relation as a set.}
\item{concurrence}{A string describing the concurrency relation.}
\item{full_result}{The Allen relation as a vector.}
\item{six_value_result}{The Allen relation for relations with distinct endpoints.}
\item{full_proportion}{The proportion of each relation in an empirical Allen relation.}
\item{six_value_proportion}{The proportion of relations with distinct endpoints in an empirical Allen relation.}
\item{concurrence_proportion}{The proportion of concurrent relations in an empirical Allen relation.}
\item{mcmc_file}{File name passed to the \code{mcmc} parameter.}
\item{application}{String passed to the \code{app} parameter.}
}
}
\description{
Summarize the relation of two phases
}
\author{
Thomas S. Dye
}
