% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_set_to_vector}
\alias{allen_set_to_vector}
\title{Convert an Allen relation set to a named vector}
\usage{
allen_set_to_vector(s)
}
\arguments{
\item{s}{An Allen relation set, a vector of single letter codes.}
}
\value{
A named result vector.
}
\description{
Set elements that are not Allen relation codes are silently ignored.
}
\author{
Thomas S. Dye
}
