% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_six_value_set}
\alias{allen_six_value_set}
\title{Allen relation set for intervals with distinct endpoints.}
\usage{
allen_six_value_set()
}
\value{
An Allen relation set
}
\description{
Return the six value Allen relation set for intervals with distinct
endpoints.
}
\author{
Thomas S. Dye
}
