% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_string_to_vector}
\alias{allen_string_to_vector}
\title{Convert a string containing Allen relation codes to a result vector}
\usage{
allen_string_to_vector(s)
}
\arguments{
\item{s}{A string with Allen relation codes}
}
\value{
A named result vector
}
\description{
A result vector is named with Allen relation codes and contains counts of
observed relations.
}
\author{
Thomas S. Dye
}
