% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allen-mcmc.R
\name{allen_table}
\alias{allen_table}
\title{Table of Allen Relations}
\usage{
allen_table(...)
}
\value{
A \code{\link{data.frame}}.
}
\description{
Table of Allen Relations
}
\author{
N. Frerebeau
}
\keyword{internal}
