% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_update_result}
\alias{allen_update_result}
\title{Update a result vector}
\usage{
allen_update_result(relation, result_vector)
}
\arguments{
\item{relation}{The relation to increment}

\item{result_vector}{The result vector to update}
}
\value{
The updated result vector
}
\description{
Increment the element of the result vector corresponding to the given
relation.
}
\author{
Thomas S. Dye
}
