% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{analyze_allen_relations}
\alias{analyze_allen_relations}
\title{Data for an analytic graphic}
\usage{
analyze_allen_relations(relation_1, relation_2, title)
}
\arguments{
\item{relation_1}{a string representation of an Allen relation}

\item{relation_2}{a string representation of an Allen relation}

\item{title}{a title for the plot}
}
\value{
A dataframe for input to \code{allen_plot}
}
\description{
Calculates the Allen composition of two relations
}
\author{
Thomas S. Dye
}
