% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchaeoPhases-defunct.R
\name{ArchaeoPhases-defunct}
\alias{ArchaeoPhases-defunct}
\alias{AgeDepth}
\alias{CreateMinMaxGroup}
\alias{CredibleInterval}
\alias{credible_interval}
\alias{DatesHiatus}
\alias{dates_hiatus}
\alias{estimate_range}
\alias{MarginalPlot}
\alias{marginal_plot}
\alias{MarginalProba}
\alias{MarginalStatistics}
\alias{marginal_statistics}
\alias{multi_marginal_statistics}
\alias{MultiCredibleInterval}
\alias{multi_credible_interval}
\alias{MultiDatesPlot}
\alias{multi_dates_plot}
\alias{MultiHPD}
\alias{multi_hpd}
\alias{MultiMarginalPlot}
\alias{multi_marginal_plot}
\alias{MultiPhasePlot}
\alias{MultiPhaseTimeRange}
\alias{MultiPhasesGap}
\alias{MultiPhasesTransition}
\alias{MultiSuccessionPlot}
\alias{OccurrencePlot}
\alias{occurrence_plot}
\alias{PhaseDurationPlot}
\alias{PhasePlot}
\alias{PhaseStatistics}
\alias{PhaseTimeRange}
\alias{PhasesGap}
\alias{phases_gap}
\alias{PhasesTransition}
\alias{SuccessionPlot}
\alias{TempoActivityPlot}
\alias{tempo_activity_plot}
\alias{TempoPlot}
\alias{tempo_plot}
\alias{undated_sample}
\title{Defunct Functions in ArchaeoPhases}
\usage{
AgeDepth(...)

CreateMinMaxGroup(...)

CredibleInterval(...)

credible_interval(...)

DatesHiatus(...)

dates_hiatus(...)

estimate_range(...)

MarginalPlot(...)

marginal_plot(...)

MarginalProba(...)

MarginalStatistics(...)

marginal_statistics(...)

multi_marginal_statistics(...)

MultiCredibleInterval(...)

multi_credible_interval(...)

MultiDatesPlot(...)

multi_dates_plot(...)

MultiHPD(...)

multi_hpd(...)

MultiMarginalPlot(...)

multi_marginal_plot(...)

MultiPhasePlot(...)

MultiPhaseTimeRange(...)

MultiPhasesGap(...)

MultiPhasesTransition(...)

MultiSuccessionPlot(...)

OccurrencePlot(...)

occurrence_plot(...)

PhaseDurationPlot(...)

PhasePlot(...)

PhaseStatistics(...)

PhaseTimeRange(...)

PhasesGap(...)

phases_gap(...)

PhasesTransition(...)

SuccessionPlot(...)

TempoActivityPlot(...)

tempo_activity_plot(...)

TempoPlot(...)

tempo_plot(...)

undated_sample(...)
}
\arguments{
\item{...}{Not used.}
}
\description{
These functions are defunct and have been replaced in \pkg{ArchaeoPhases}
2.0.
}
\details{
See the changelog for details:
\code{news(Version >= "2.0", package = "ArchaeoPhases")}.
}
\keyword{internal}
