% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchaeoPhases-deprecated.R
\name{ArchaeoPhases-deprecated}
\alias{ArchaeoPhases-deprecated}
\title{Deprecated Functions in ArchaeoPhases}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\keyword{internal}
