% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R
\docType{methods}
\name{as.array}
\alias{as.array}
\alias{as.matrix,MCMC-method}
\alias{as.array,PhasesMCMC-method}
\title{Coerce to an Array}
\usage{
\S4method{as.matrix}{MCMC}(x, ...)

\S4method{as.array}{PhasesMCMC}(x, ...)
}
\arguments{
\item{x}{An \R object.}

\item{...}{Currently not used.}
}
\value{
An \code{\link{array}} or a \code{\link{matrix}}.
}
\description{
Coerce to an Array
}
\note{
June 2026: aperm.default() now copy attributes, this change how apply()
works on classed objects. The simplest "fix" is to add an aperm() S3 method
for the class which drops attributes as aperm.default() previously did.
Alternatively, one could also consider providing as.matrix()/as.array()
methods which change to standard matrix/array.
}
\author{
N. Frerebeau
}
\keyword{internal}
