% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{calendar}
\alias{calendar}
\alias{BP_to_BCAD}
\alias{BP_to_BCAD-method}
\alias{BCAD_to_BP}
\alias{BCAD_to_BP-method}
\alias{is_BP}
\alias{is_BP-method}
\alias{is_CE}
\alias{is_CE-method}
\alias{is_BP,CumulativeEvents-method}
\alias{is_CE,CumulativeEvents-method}
\alias{BP_to_BCAD,numeric-method}
\alias{BP_to_BCAD,matrix-method}
\alias{BP_to_BCAD,array-method}
\alias{BP_to_BCAD,MCMC-method}
\alias{BP_to_BCAD,PhasesMCMC-method}
\alias{BP_to_BCAD,CumulativeEvents-method}
\alias{BCAD_to_BP,numeric-method}
\alias{BCAD_to_BP,matrix-method}
\alias{BCAD_to_BP,array-method}
\alias{BCAD_to_BP,MCMC-method}
\alias{BCAD_to_BP,PhasesMCMC-method}
\alias{BCAD_to_BP,CumulativeEvents-method}
\title{Convert Time Scale}
\usage{
BP_to_BCAD(object)

BCAD_to_BP(object)

is_BP(object)

is_CE(object)

\S4method{is_BP}{CumulativeEvents}(object)

\S4method{is_CE}{CumulativeEvents}(object)

\S4method{BP_to_BCAD}{numeric}(object)

\S4method{BP_to_BCAD}{matrix}(object)

\S4method{BP_to_BCAD}{array}(object)

\S4method{BP_to_BCAD}{MCMC}(object)

\S4method{BP_to_BCAD}{PhasesMCMC}(object)

\S4method{BP_to_BCAD}{CumulativeEvents}(object)

\S4method{BCAD_to_BP}{numeric}(object)

\S4method{BCAD_to_BP}{matrix}(object)

\S4method{BCAD_to_BP}{array}(object)

\S4method{BCAD_to_BP}{MCMC}(object)

\S4method{BCAD_to_BP}{PhasesMCMC}(object)

\S4method{BCAD_to_BP}{CumulativeEvents}(object)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector or an \code{\linkS4class{MCMC}} object.}
}
\value{
\itemize{
\item \code{is_BP} and \code{is_CE} return a \code{\link{logical}} scalar.
}

An object of the same sort as \code{object} with a new time scale.
}
\description{
Converts between BP and BC/AD time scales.
}
\note{
There is no year \eqn{0} in BCE/CE scale.
}
\author{
N. Frerebeau
}
\concept{time scales}
