% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/interval.R
\docType{methods}
\name{ci}
\alias{ci}
\alias{interval_credible}
\alias{interval_credible-method}
\alias{interval_credible,numeric-method}
\alias{interval_credible,MCMC-method}
\title{Bayesian Credible Interval}
\usage{
interval_credible(object, ...)

\S4method{interval_credible}{numeric}(object, level = 0.95)

\S4method{interval_credible}{MCMC}(object, level = 0.95)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector or an \code{\linkS4class{MCMC}} object containing the
output of the MCMC algorithm for the parameter.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\value{
A two columns \code{numeric} \code{\link{matrix}} giving the lower and upper boundaries of
the credible interval.
}
\description{
Computes the shortest credible interval of the output of the MCMC algorithm
for a single parameter.
}
\details{
A \eqn{(100 \times level)}{(100 * level)}\\% credible interval is an interval
that keeps \eqn{N \times (1 - level)}{N * (1 - level)} elements of the
sample outside the interval.

The \eqn{(100 \times level)}{(100 * level)}\\% credible interval is the
shortest of all those intervals.
}
\examples{
## Coerce to MCMC
events <- as_mcmc(Events, iteration = 1)
events <- events[1:10000, ]

## Credible interval
interval_credible(events, level = 0.95)

## HPD regions
interval_hpd(events, level = 0.95)
}
\seealso{
Other statistics: 
\code{\link{hpdi}},
\code{\link{summary}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{statistics}
