% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R
\docType{methods}
\name{coerce}
\alias{coerce}
\alias{as_mcmc}
\alias{as_mcmc-method}
\alias{as_events}
\alias{as_events-method}
\alias{as_mcmc,matrix-method}
\alias{as_events,matrix-method}
\alias{as_mcmc,data.frame-method}
\alias{as_events,data.frame-method}
\title{Coerce}
\usage{
as_mcmc(from, ...)

as_events(from, ...)

\S4method{as_mcmc}{matrix}(from, BP = FALSE, iteration = NULL)

\S4method{as_events}{matrix}(from, BP = FALSE, iteration = NULL)

\S4method{as_mcmc}{data.frame}(from, BP = FALSE, iteration = NULL)

\S4method{as_events}{data.frame}(from, BP = FALSE, iteration = NULL)
}
\arguments{
\item{from}{from An object to be coerced.}

\item{...}{Currently not used.}

\item{BP}{A \code{\link{logical}} scalar: should the data be converted from BP to
BC/AD? This should not be \code{TRUE} unless you change the default settings in
'OxCal' or 'ChronoModel'.}

\item{iteration}{An \code{\link{integer}} specifying the index of the iteration column
to be removed.}
}
\value{
An \code{\linkS4class{MCMC}} object.
}
\description{
Coerce
}
\examples{
## Coerce to MCMC
eve <- as_mcmc(events, iteration = 1)
}
\seealso{
Other read methods: 
\code{\link{check}},
\code{\link{coda}},
\code{\link{read}}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{read methods}
