% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenComposition.R
\name{composition.lookup.table}
\alias{composition.lookup.table}
\title{Construct an Allen composition lookup table}
\usage{
composition.lookup.table()
}
\description{
Construct an Allen composition lookup table
}
\author{
Thomas S. Dye
}
