% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{composition_lookup_table}
\alias{composition_lookup_table}
\title{Construct an Allen composition lookup table}
\usage{
composition_lookup_table()
}
\description{
Construct an Allen composition lookup table
}
\author{
Thomas S. Dye
}
