% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenVector.R
\name{convert.to.result.vector}
\alias{convert.to.result.vector}
\title{Convert to a result vector}
\usage{
convert.to.result.vector(obj)
}
\arguments{
\item{obj}{Object to convert to a result vector.}
}
\value{
A result vector or an error if obj is not an Allen set.
}
\description{
Expects an Allen set represented as a string, vector, or result vector
and converts, if necessary, to a result vector.
}
\author{
Thomas S. Dye
}
