% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/interval.R
\docType{methods}
\name{credible}
\alias{credible}
\alias{credible-method}
\alias{credible,numeric-method}
\alias{credible,MCMC-method}
\title{Bayesian Credible Interval}
\usage{
credible(object, ...)

\S4method{credible}{numeric}(object, level = 0.95, CE = TRUE, ...)

\S4method{credible}{MCMC}(object, level = 0.95, ...)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector or an \code{\linkS4class{MCMC}} object containing the
output of the MCMC algorithm for the parameter.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}

\item{CE}{A \code{\link{logical}} scalar: are the data expressed in BC/AD years?}
}
\value{
A \code{\link{list}} of \code{numeric} \code{\link{matrix}} giving the lower and upper boundaries of
the credible interval and associated probabilities.
}
\description{
Computes the shortest credible interval of the output of the MCMC algorithm
for a single parameter.
}
\details{
A \eqn{(100 \times level)}{(100 * level)} \% credible interval is an interval
that keeps \eqn{N \times (1 - level)}{N * (1 - level)} elements of the
sample outside the interval.

The \eqn{(100 \times level)}{(100 * level)} \% credible interval is the
shortest of all those intervals.

For instance, the 95\% credible interval is the central portion of the
posterior distribution that contains 95\% of the values.
}
\examples{
## Coerce to MCMC
eve <- as_events(events, calendar = "CE", iteration = 1)
eve <- eve[1:10000, ]

## CE
credible(eve, level = 0.95) # Credible interval
hpdi(eve, level = 0.95) # HPD interval

## BP
eve_BP <- CE_to_BP(eve)
credible(eve_BP, level = 0.95) # Credible interval
hpdi(eve_BP, level = 0.95) # HPD interval
}
\seealso{
Other statistics: 
\code{\link{hpdi}()},
\code{\link{sensitivity}()},
\code{\link{summary}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{statistics}
