% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CumulativeEvents}
\alias{CumulativeEvents}
\alias{CumulativeEvents-class}
\alias{.CumulativeEvents}
\title{Cumulative Events}
\description{
An S4 class to store the result of a \code{\link{tempo}} plot.
}
\section{Slots}{

\describe{
\item{\code{level}}{A length-one \code{\link{numeric}} vector giving the confidence level.}

\item{\code{calendar}}{A \code{\link{character}} string specifying the chronological scale
(either "\code{BP}" or "\code{BCAD}").}

\item{\code{hash}}{A \code{\link{character}} string giving the a cryptographical hash of the
original data file.}
}}

\seealso{
Other class: 
\code{\link{ActivityEvents}},
\code{\link{MCMC}},
\code{\link{OccurrenceEvents}}
}
\author{
N. Frerebeau
}
\concept{class}
