% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSet.R
\name{ensure.allen.set.vector}
\alias{ensure.allen.set.vector}
\title{Ensure an Allen set is represented as a vector of single character strings}
\usage{
ensure.allen.set.vector(obj)
}
\arguments{
\item{obj}{An Allen set represented as a string, a set vector, or a
result vector.}
}
\value{
An Allen set represented as a set vector.
}
\description{
Expects a string, set vector, or result vector and will stop with an
error if something else is encountered.
}
\author{
Thomas S. Dye
}
