% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dates.R
\docType{methods}
\name{gap}
\alias{gap}
\alias{date_hiatus}
\alias{date_hiatus-method}
\alias{date_hiatus,numeric,numeric-method}
\title{Hiatus Between Two Dates}
\usage{
date_hiatus(x, y, ...)

\S4method{date_hiatus}{numeric,numeric}(x, y, level = 0.95)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\description{
Tests for the existence of a hiatus between two parameters.
}
\details{
\code{date_hiatus()} finds if a gap exists between two dates and returns the
longest interval that satisfies:
\eqn{P(x < HiatusInf < HiatusSup < y | M) = level}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{dates}
