% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/interval.R
\docType{methods}
\name{hpdi}
\alias{hpdi}
\alias{interval_hpd}
\alias{interval_hpd-method}
\alias{interval_hpd,MCMC-method}
\title{Bayesian HPD Regions}
\usage{
interval_hpd(object, ...)

\S4method{interval_hpd}{MCMC}(object, level = 0.95, ...)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector or an \code{\linkS4class{MCMC}} object containing the
output of the MCMC algorithm for the parameter.}

\item{...}{Extra arguments to be passed to \code{\link[hdrcde:hdr]{hdrcde::hdr()}}.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\value{
A \code{\link{list}} of two columns \code{numeric} \code{\link{matrix}} giving the lower and upper
boundaries of the HPD interval.
}
\description{
Bayesian HPD Regions
}
\examples{
## Coerce to MCMC
eve <- as_mcmc(events, iteration = 1)
eve <- eve[1:10000, ]

## Credible interval
interval_credible(eve, level = 0.95)

## HPD regions
interval_hpd(eve, level = 0.95)
}
\references{
Hyndman, R. J. (1996). Computing and graphing highest density regions.
\emph{American Statistician}, 50: 120-126. \doi{10.2307/2684423}.
}
\seealso{
\code{\link[hdrcde:hdr]{hdrcde::hdr()}}

Other statistics: 
\code{\link{ci}},
\code{\link{summary}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{statistics}
