% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenPredicates.R
\name{is.illustration.vector}
\alias{is.illustration.vector}
\title{Test if an object is an illustration vector}
\usage{
is.illustration.vector(obj)
}
\arguments{
\item{obj}{An object to test}
}
\value{
Boolean, TRUE if obj is an illustration vector, FALSE otherwise.
}
\description{
An illustration vector is a vector whose elements are all
zeros and ones.  It can be used to illustrate an Allen set.
}
\author{
Thomas S. Dye
}
