% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenSupport.R
\name{is.odd}
\alias{is.odd}
\title{Check if an argument is odd.}
\usage{
is.odd(x)
}
\arguments{
\item{x}{An integer.}
}
\value{
Boolean
}
\description{
Utility function that checks if its numeric argument is odd.
Stops with an error if the argument is non-numeric.
}
