% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenPredicates.R
\name{is.result.vector}
\alias{is.result.vector}
\title{Test whether an object is a result vector}
\usage{
is.result.vector(obj)
}
\arguments{
\item{obj}{An object to test}
}
\value{
Boolean, TRUE if obj is a result vector, FALSE otherwise.
}
\description{
Checks vector length, names, and class
}
