% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenPredicates.R
\name{is.set.string}
\alias{is.set.string}
\title{Test if an object is a set string}
\usage{
is.set.string(obj)
}
\arguments{
\item{obj}{An object to test}
}
\value{
Boolean, TRUE if obj is a set string, FALSE otherwise.
}
\description{
Checks for mode 'character', length of 1, and nchar <= 13
}
\author{
Thomas S. Dye
}
