% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenPredicates.R
\name{is.set.vector}
\alias{is.set.vector}
\title{Test if an object is a set vector}
\usage{
is.set.vector(obj)
}
\arguments{
\item{obj}{An object to test}
}
\value{
Boolean, TRUE if obj is a set vector, FALSE otherwise.
}
\description{
Checks for mode 'character', length less than 13.
Note: this predicate is a (very) partial implementation.
}
\author{
Thomas S. Dye
}
