% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportCSV.R
\name{is.url}
\alias{is.url}
\title{Check if string is a URL}
\usage{
is.url(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
Returns a logical evalution as to whether a string is a URL.
}
\description{
Uses a regex approach to check if a string is a URL.  This approach is faster
than \code{\link[RCurl]{url.exists}} but does do the actual verification.
}
\details{
Retrieved from:
https://github.com/trinker/reports/blob/master/R/is.url.R
}
