% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/test.R
\docType{methods}
\name{lapse}
\alias{lapse}
\alias{lapse-method}
\alias{lapse,numeric,numeric-method}
\alias{lapse,EventsMCMC,missing-method}
\alias{lapse,EventsMCMC-method}
\title{Hiatus Between Two Dates}
\usage{
lapse(x, y, ...)

\S4method{lapse}{numeric,numeric}(x, y, level = 0.95)

\S4method{lapse}{EventsMCMC,missing}(x, y)
}
\arguments{
\item{x}{A \code{\link{numeric}} vector giving the output of the MCMC algorithm for the
first parameter, or an \code{\linkS4class{EventsMCMC}} object.}

\item{y}{A \code{\link{numeric}} vector giving the output of the MCMC algorithm for the
second parameter.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\description{
Tests for the existence of a hiatus between two parameters.
}
\details{
Finds if a gap exists between two dates and returns the longest interval
that satisfies: \eqn{P(x < HiatusInf < HiatusSup < y | M) = level}
}
\section{Methods (by class)}{
\itemize{
\item \code{x = numeric,y = numeric}: Returns a length-three \code{\link{numeric}} vector (upper and upper
boundaries, and hiatus duration, if any).

\item \code{x = EventsMCMC,y = missing}: Returns a \code{\link{numeric}} matrix of hiatus durations.
}}

\examples{
## Coerce to MCMC
eve <- as_events(events, iteration = 1)
eve <- eve[1:10000, ]

## Test for anteriority
older(eve)

## Test for hiatus
lapse(eve)
}
\seealso{
Other tests: 
\code{\link{apportion}()},
\code{\link{older}()}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{tests}
