% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmc_events}
\alias{mcmc_events}
\title{Events}
\format{
A \code{\link{data.frame}} with 30,000 rows and 5 variables:
\describe{
\item{\code{iter}}{Iteration of the MCMC algorithm.}
\item{\code{E1}}{Information on event 1.}
\item{\code{E2}}{Information on event 2.}
\item{\code{E3}}{Information on event 3.}
\item{\code{E4}}{Information on event 4.}
}
}
\usage{
mcmc_events
}
\description{
A data set containing information on the ages of four dated events.
}
\seealso{
Other datasets:
\code{\link{mcmc_phases}}
}
\concept{datasets}
\keyword{datasets}
