% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmc_phases}
\alias{mcmc_phases}
\title{Phases}
\format{
A \code{\link{data.frame}} with 30,000 rows and 5 variables:
\describe{
\item{\code{iter}}{Iteration of the MCMC algorithm.}
\item{\code{P2_alpha}}{Start date of Phase 2.}
\item{\code{P2_beta}}{End date of Phase 2.}
\item{\code{P1_alpha}}{Start date of Phase 1.}
\item{\code{P1_beta}}{End date of Phase 1.}
}
}
\usage{
mcmc_phases
}
\description{
A data set containing information on the start and end dates of two phases.
}
\seealso{
Other datasets:
\code{\link{mcmc_events}}
}
\concept{datasets}
\keyword{datasets}
