% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/multiplot.R
\docType{methods}
\name{multiplot}
\alias{multiplot}
\alias{multiplot-method}
\alias{multiplot,CumulativeEvents-method}
\alias{multiplot,ActivityEvents-method}
\title{Multiple Plots}
\usage{
multiplot(...)

\S4method{multiplot}{CumulativeEvents}(...)

\S4method{multiplot}{ActivityEvents}(...)
}
\arguments{
\item{...}{Any object of the same class.}
}
\value{
\code{multiplot()} is called it for its side-effects: it results in a graphic
being displayed.
}
\description{
Multiple Plots
}
\seealso{
Other plot methods: 
\code{\link{plot}()}
}
\author{
N. Frerebeau
}
\concept{plot methods}
