% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{mutator}
\alias{mutator}
\alias{get}
\alias{set}
\alias{names,MCMC-method}
\alias{names<-,MCMC-method}
\alias{names,PhasesMCMC-method}
\alias{names<-,PhasesMCMC-method}
\title{Get or Set Parts of an Object}
\usage{
\S4method{names}{MCMC}(x)

\S4method{names}{MCMC}(x) <- value

\S4method{names}{PhasesMCMC}(x)

\S4method{names}{PhasesMCMC}(x) <- value
}
\arguments{
\item{x}{An object from which to get or set element(s).}

\item{value}{A possible value for the element(s) of \code{x}.}
}
\value{
An object of the same sort as \code{x} with the new values assigned.
}
\description{
Getters and setters to extract or replace parts of an object.
}
\seealso{
Other mutators: 
\code{\link{as_coda}()},
\code{\link{bind}},
\code{\link{sort.list}()},
\code{\link{sort}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
