% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{new_allen_illustrative}
\alias{new_allen_illustrative}
\title{Constructor for illustrative Allen set object}
\usage{
new_allen_illustrative(x = data.frame())
}
\arguments{
\item{x}{A data frame with the Allen set.}
}
\value{
An \code{allen_set_illustrative} object that inherits from \code{tbl_df}.
}
\description{
Object to be returned by functions that create illustrative Allen sets.
}
\author{
Thomas S. Dye, \email{tsd@tsdye.online}
}
