% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Classes.R
\name{new_archaeophases_mcmc}
\alias{new_archaeophases_mcmc}
\title{Constructor for \code{archaeophases_mcmc} object}
\usage{
new_archaeophases_mcmc(x = list(), call = language(), hash = character())
}
\arguments{
\item{x}{A data frame with the data from the csv file.}

\item{call}{How the function was called.}

\item{hash}{A SHA256 hash of the csv file.}
}
\value{
An \code{archaeophases_mcmc} object that inherits from \code{tbl_df}.
}
\description{
Object to be returned by functions that read MCMC data from csv files.
}
\details{
The SHA256 hash should be secure against intentional and unintentional
alterations of the MCMC csv file.
}
\seealso{
\code{\link{read_chronomodel}}

\code{\link{read_bcal}}

\code{\link{read_oxcal}}
}
\author{
Thomas S. Dye, \email{tsd@tsdye.online}
}
