% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Classes.R
\name{original_file.archaeophases_mcmc}
\alias{original_file.archaeophases_mcmc}
\title{Check for an original mcmc file}
\usage{
\method{original_file}{archaeophases_mcmc}(x, file = NULL)
}
\arguments{
\item{x}{An \code{archaeophases_mcmc} object.}

\item{file}{Either a path to a CSV file, a connection,
or the value \code{clipboard()} to read from the system clipboard.
The CSV \code{file} can be compressed or plain.}
}
\value{
A boolean, \code{TRUE} if the files match, \code{FALSE}
otherwise.
}
\description{
Checks whether or not a file is identical to the one used to create
an \code{archaeophases_mcmc} object.
}
\details{
If called with a single argument, checks the file indicated by
the \code{file_path} attribute.
}
\author{
Thomas S. Dye, \email{tsd@tsdye.online}
}
