% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxc}
\alias{oxc}
\title{oxc}
\format{
A data frame with 1000 rows and 2 variables:
\describe{
\item{foo-early}{date oo-early}
\item{foo-late}{date foo-late}

}
}
\usage{
oxc
}
\description{
A data set containing information on the ages of two events.
see the vignette Reproductibility for more details
}
\keyword{datasets}
