% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{oxcal_calibrate}
\alias{oxcal_calibrate}
\title{14C Calibration with OxCal}
\usage{
oxcal_calibrate(names, dates, errors, curve = "IntCal20")
}
\arguments{
\item{names}{A \code{\link{character}} vector specifying the names of the dates (e.g.
laboratory codes).}

\item{dates}{A \code{\link{numeric}} vector giving the BP dates to be calibrated.}

\item{errors}{A \code{\link{numeric}} vector giving the standard deviation of the dates
to be calibrated.}

\item{curve}{A \code{\link{character}} string specifying the calibration curve to be
used.}
}
\value{
A \code{\link{list}} with the following elements:
\describe{
\item{\code{ocd}}{A \code{list} holding the ranges, probability distributions, etc.
for each parameter.}
\item{\code{model}}{A \code{list} containing information about the model.}
\item{\code{calib}}{A \code{list} containing information about the calibration
curve.}
}
}
\description{
14C Calibration with OxCal
}
\examples{
\dontrun{
## Download OxCal
oxcal_setup()

## Calibrate 14C dates
cal <- oxcal_calibrate(
  names = c("X","Y"),
  dates = c(5000, 4500),
  errors = c(45, 35)
)
plot(cal, curve = FALSE)
plot(cal, curve = TRUE)

## Custom script
scr <- 'Plot()
 {
  Sequence("Sequence1")
  {
   Boundary("Begin");
   Phase("Phase1")
   {
    R_Date("Lab-1",5000,25);
    R_Date("Lab-2",4900,37);
   };
   Boundary("Between");
   Phase("Phase2")
   {
    R_Date("Lab-3",4800,43);
   };
   Boundary("End");
  };
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)

plot(res, curve = FALSE)
plot(res, curve = TRUE)

as.data.frame(res)
}
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_download}()},
\code{\link{oxcal_execute}()},
\code{\link{oxcal_parse}()},
\code{\link{oxcal_setup}()}
}
\author{
N. Frerebeau
}
\concept{OxCal tools}
