% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{oxcal_download}
\alias{oxcal_download}
\title{Download OxCal}
\usage{
oxcal_download(path = NULL)
}
\arguments{
\item{path}{A \code{\link{character}} string specifying the directory to extract files
to. It will be created if necessary (see \code{\link[utils:unzip]{utils::unzip()}}).}
}
\value{
Invisibly returns \code{path}.
}
\description{
Download OxCal
}
\note{
Adapted from \code{\link[oxcAAR:quickSetupOxcal]{oxcAAR::quickSetupOxcal()}}.
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_calibrate}()},
\code{\link{oxcal_execute}()},
\code{\link{oxcal_setup}()}
}
\author{
N. Frerebeau
}
\concept{OxCal tools}
