% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{oxcal_parse}
\alias{oxcal_parse}
\title{Read and Parse OxCal Output}
\usage{
oxcal_parse(file)
}
\arguments{
\item{file}{A \code{\link{character}} string naming a JavaScript file which the data
are to be read from (typically returned by \code{\link[=oxcal_execute]{oxcal_execute()}}).}
}
\description{
Read and Parse OxCal Output
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_calibrate}()},
\code{\link{oxcal_download}()},
\code{\link{oxcal_execute}()},
\code{\link{oxcal_setup}()}
}
\author{
N. Frerebeau
}
\concept{OxCal tools}
