% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{oxcal_setup}
\alias{oxcal_setup}
\title{Setup OxCal}
\usage{
oxcal_setup(path = NULL, os = NULL, ask = TRUE)
}
\arguments{
\item{path}{A \code{\link{character}} string specifying the directory where to find the
OxCal executable (or to extract OxCal files to).}

\item{os}{A \code{\link{character}} string specifying the operating system of the
workstation. It must be one of "\code{Linux}", "\code{Windows}" or "\code{Darwin}".
If \code{NULL} (the default), the operating system will be determined
automatically (see \code{\link[=Sys.info]{Sys.info()}}).}

\item{ask}{A \code{\link{logical}} scalar: if OxCal is not installed, should the user
be asked to download it?
If \code{FALSE} and the OxCal executable cannot be found, will raise an error.
Only used if \R is being used interactively.}
}
\value{
Invisibly returns the path to the OxCal executable.
}
\description{
Setup OxCal
}
\details{
Downloads the latest version of Oxcal (if needed) and sets the executable
path correctly.
}
\note{
Adapted from \code{\link[oxcAAR:quickSetupOxcal]{oxcAAR::quickSetupOxcal()}}.
}
\examples{
\dontrun{
## Download OxCal
oxcal_setup()

## Calibrate 14C dates
cal <- oxcal_calibrate(
  names = c("X","Y"),
  dates = c(5000, 4500),
  errors = c(45, 35)
)
plot(cal)

## Custom script
scr <- 'Plot()
 {
  R_Date("X",5000,25);
  R_Date("Y",2500,50);
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)
res
}
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_calibrate}()},
\code{\link{oxcal_download}()},
\code{\link{oxcal_execute}()},
\code{\link{oxcal_parse}()}
}
\author{
C. Schmid, N. Frerebeau
}
\concept{OxCal tools}
