% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/phases.R
\docType{methods}
\name{phase}
\alias{phase}
\alias{phase-method}
\alias{get_order}
\alias{get_order-method}
\alias{set_order<-}
\alias{set_order-method}
\alias{as_ordered}
\alias{as_ordered-method}
\alias{is_ordered}
\alias{is_ordered-method}
\alias{phase,MCMC,missing-method}
\alias{phase,MCMC,list-method}
\alias{set_order<-,PhasesMCMC,character-method}
\alias{set_order,PhasesMCMC,character-method}
\alias{get_order,PhasesMCMC-method}
\alias{as_ordered,PhasesMCMC-method}
\alias{is_ordered,PhasesMCMC-method}
\title{Phase}
\usage{
phase(x, groups, ...)

get_order(x, value)

set_order(x) <- value

as_ordered(x)

is_ordered(x)

\S4method{phase}{MCMC,missing}(x)

\S4method{phase}{MCMC,list}(x, groups, ordered = FALSE)

\S4method{set_order}{PhasesMCMC,character}(x) <- value

\S4method{get_order}{PhasesMCMC}(x)

\S4method{as_ordered}{PhasesMCMC}(x)

\S4method{is_ordered}{PhasesMCMC}(x)
}
\arguments{
\item{x}{An \code{\linkS4class{MCMC}} or a \code{\linkS4class{PhasesMCMC}} object.}

\item{groups}{A \code{\link{list}}.}

\item{...}{Currently not used.}

\item{value}{A possible value for the element(s) of \code{x}.}

\item{ordered}{A \code{\link{logical}} scalar: should the \code{groups} be regarded as
ordered (in the order given)?}
}
\value{
A \code{\linkS4class{PhasesMCMC}} object.
}
\description{
Constructs the minimum and maximum for a group of events (phase).
}
\examples{
## Coerce to MCMC objects
## Events
(eve <- as_events(events[1:10000, ], iteration = 1))
## Phases
(pha <- as_phases(phases[1:10000, ], start = c(1, 3), iteration = 1))

## Compute phases
## Compute min-max range for all chains
pha1 <- phase(eve)

## Compute min-max range by group
## Unless otherwise specified, the phases are assumed to be unordered
pha2 <- phase(eve, groups = list(A = c(1, 3), B = c(2, 4)))

all(pha == pha2)

## Set chronological order
## (from the oldest to the youngest phase)
set_order(pha2) <- c("B", "A")
get_order(pha2)
is_ordered(pha2)
}
\seealso{
Other phases: 
\code{\link{boundaries}()},
\code{\link{duration}()},
\code{\link{hiatus}()},
\code{\link{transition}()}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{phases}
