% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PhasesMCMC}
\alias{PhasesMCMC}
\alias{PhasesMCMC-class}
\alias{.PhasesMCMC}
\title{MCMC Phases}
\description{
An S4 class to represent the output of a MCMC algorithm.
}
\section{Slots}{

\describe{
\item{\code{start}}{An \code{\link{integer}} vector containing the column number corresponding
to the minimum of each phase.}

\item{\code{end}}{An \code{\link{integer}} vector containing the column number corresponding to
the maximum of each phase (set in the same order as in \code{start}).}

\item{\code{phases}}{A \code{\link{factor}} vector specifying the name of the phases.}

\item{\code{calendar}}{A \code{\link{character}} string specifying the chronological scale
(either "\code{BP}" or "\code{BCAD}").}

\item{\code{hash}}{A \code{\link{character}} string giving the cryptographical hash of the
original data file.}
}}

\note{
These classes inherit from \code{\link{array}}.
}
\section{Coerce}{

In the code snippets below, \code{x} is a \code{PhasesMCMC} object.
\describe{
\item{\code{as.list(x)}}{Coerces to a \code{\link{list}}.}
}
}

\section{Subset}{

In the code snippets below, \code{x} is a \code{PhasesMCMC} object.
\describe{
\item{\code{x[[i]]}}{Extracts a single phase (two chains) selected by subscript
\code{i}. \code{i} is a length-one \code{\link{numeric}} or \code{\link{character}} vector.}
}
}

\seealso{
Other class: 
\code{\link{ActivityEvents}},
\code{\link{CumulativeEvents}},
\code{\link{MCMC}},
\code{\link{OccurrenceEvents}}
}
\author{
N. Frerebeau
}
\concept{class}
