% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot-method}
\alias{plot,MCMC,missing-method}
\alias{plot,PhasesMCMC,missing-method}
\title{Plot}
\usage{
\S4method{plot}{MCMC,missing}(
  x,
  density = TRUE,
  interval = NULL,
  level = 0.95,
  decreasing = TRUE,
  main = NULL,
  sub = NULL,
  ann = graphics::par("ann"),
  axes = TRUE,
  frame.plot = FALSE,
  panel.first = NULL,
  panel.last = NULL,
  ...
)

\S4method{plot}{PhasesMCMC,missing}(
  x,
  density = TRUE,
  boundaries = TRUE,
  range = NULL,
  level = 0.95,
  decreasing = TRUE,
  main = NULL,
  sub = NULL,
  ann = graphics::par("ann"),
  axes = TRUE,
  frame.plot = FALSE,
  panel.first = NULL,
  panel.last = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{\linkS4class{MCMC}} or a \code{\linkS4class{PhasesMCMC}} object.}

\item{density}{A \code{\link{logical}} scalar: should estimated density be plotted?}

\item{interval}{A \code{\link{character}} string specifying the confidence interval to
be drawn. It must be one of "\code{credible}" (credible interval) or "\code{hpdi}"
(highest posterior density interval). Any unambiguous substring can be
given. If \code{NULL} (the default) no interval is computed.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}

\item{decreasing}{A \code{\link{logical}} scalar: should the sort order be decreasing?}

\item{main}{A \code{\link{character}} string giving a main title for the plot.}

\item{sub}{A \code{\link{character}} string giving a subtitle for the plot.}

\item{ann}{A \code{\link{logical}} scalar: should the default annotation (title and x,
y and z axis labels) appear on the plot?}

\item{axes}{A \code{\link{logical}} scalar: should axes be drawn on the plot?}

\item{frame.plot}{A \code{\link{logical}} scalar: should a box be drawn around the
plot?}

\item{panel.first}{An an \code{expression} to be evaluated after the plot axes are
set up but before any plotting takes place. This can be useful for drawing
background grids.}

\item{panel.last}{An \code{expression} to be evaluated after plotting has taken
place but before the axes, title and box are added.}

\item{...}{Extra parameters to be passed to \code{\link[stats:density]{stats::density()}}.}

\item{boundaries}{A \code{\link{logical}} scalar: should phase time range be plotted
(see \code{\link[=boundaries]{boundaries()}})?}

\item{range}{A \code{\link{character}} string specifying the additional time range to
be displayed. It must be one of "\code{hiatus}" or "\code{transition}". If \code{NULL} (the
default), no additional time ranges are displayed.}
}
\value{
\code{plot()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}
\description{
Plot
}
\section{Methods (by class)}{
\itemize{
\item \code{plot(x = MCMC, y = missing)}: Plots of credible intervals or HPD regions of a series of
events.

\item \code{plot(x = PhasesMCMC, y = missing)}: Plots the characteristics of a group of events (phase).

}}
\examples{
## Coerce to MCMC
eve <- as_events(events, calendar = "CE", iteration = 1)

## BP
eve_BP <- CE_to_BP(eve)
summary(eve_BP)

## CE
eve_CE <- BP_to_CE(eve_BP)
summary(eve_CE)

## Plot events
plot(eve_CE, interval = "credible", level = 0.68)
plot(eve_BP, interval = "hpdi", level = 0.68)
plot(eve_BP[[1]], interval = "hpdi")

## Compute phases
pha <- phase(eve, groups = list(B = c(2, 4), A = c(1, 3)))
summary(pha)

## Plot phases
plot(pha)
plot(pha, range = "hiatus")
plot(pha, range = "transition")
}
\seealso{
\code{\link[stats:density]{stats::density()}}

Other plot methods: 
\code{\link{multiplot}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{plot methods}
