% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenPlot.R
\name{plot_lattice}
\alias{plot_lattice}
\title{Plot a Nökel lattice}
\usage{
plot_lattice(mcmc, phases_1, phases_2, app = "bcal", quiet = "partial", ...)
}
\arguments{
\item{mcmc}{path to a csv file with MCMC output}

\item{phases_1}{list of three-element vectors, the first element of which is
the phase name and the second and third elements are the column indices
for the beginning and end of the phase, respectively}

\item{phases_2}{list of three-element vectors, the first element of which is
the phase name and the second and third elements are the column indices
for the beginning and end of the phase, respectively}

\item{app}{one of 'bcal', 'oxcal', or 'chronomodel' to specify which
Bayesian calibration application produced the MCMC output}

\item{quiet}{One of 'no' to allow messages and warnings,
'partial' (default) to suppress messages and allow warnings, or 'yes'
to suppress messages and warnings.}

\item{...}{Named arguments to be passed on to \code{allen_plot()}.}
}
\value{
A layout_tbl_graph object.
}
\description{
Reads MCMC output and plots a Nökel lattice.
}
\examples{

# Plot to the R graphics device
# plot_lattice()

# Save layout_tbl_graph object to a variable
# then plot to the R graphics device

# foo <- plot_lattice()
# foo

}
\author{
Thomas S. Dye
}
