% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/proxy.R, R/summary.R
\docType{methods}
\name{proxy}
\alias{proxy}
\alias{proxy-method}
\alias{proxy,numeric-method}
\alias{autoplot,ProxyRecord-method}
\alias{plot,ProxyRecord,missing-method}
\alias{summary,ProxyRecord-method}
\title{Layer-Counted Proxy Records Uncertainties}
\usage{
proxy(depth, ...)

\S4method{proxy}{numeric}(
  depth,
  proxy,
  proxy_error,
  time,
  time_error,
  from = min(time),
  to = max(time),
  grid = NULL,
  resolution = NULL,
  calendar = c("BP", "CE", "b2k"),
  density = !samples,
  samples = TRUE,
  n = 100
)

\S4method{autoplot}{ProxyRecord}(object, ..., raw = FALSE, IQR = TRUE)

\S4method{plot}{ProxyRecord,missing}(x, raw = FALSE, IQR = TRUE, ...)

\S4method{summary}{ProxyRecord}(object, level = 0.95)
}
\arguments{
\item{depth}{A \code{\link{numeric}} vector giving the depth at which proxy values and
calendar ages were measured.}

\item{...}{Currently not used.}

\item{proxy}{A \code{\link{numeric}} vector giving the proxy values.}

\item{proxy_error}{A \code{\link{numeric}} vector giving the proxy uncertainties.}

\item{time}{A \code{\link{numeric}} vector giving the calendar ages.}

\item{time_error}{A \code{\link{numeric}} vector giving the calendar age uncertainties.}

\item{from}{A length-one \code{\link{numeric}} vector specifying the starting value of
the temporal sequence at which densities are to be estimated.}

\item{to}{A length-one \code{\link{numeric}} vector specifying the end value of the
temporal sequence at which densities are to be estimated.}

\item{grid}{A length-one \code{\link{numeric}} vector specifying the step size (in
units of \code{proxy}) at which proxy records densities are to be estimated.
If \code{NULL} (the default), equally spaced points will be used (according to
\code{options("chronos.grid")}).}

\item{resolution}{A length-one \code{\link{numeric}} vector specifying the temporal
resolution (in units of \code{time}) at which densities are to be estimated.
If \code{NULL} (the default), the smallest temporal difference between two
consecutive records will be used.}

\item{calendar}{A \code{\link{character}} string specifying the chronological scale
It must be one of "\code{BP}" (the default), "\code{CE}" or "\code{b2k}".}

\item{density}{A \code{\link{logical}} scalar: should densities be kept in the results?}

\item{samples}{A \code{\link{logical}} scalar: should samples be returned?
Defaults to \code{TRUE}.}

\item{n}{An \code{\link{integer}} specifying the number of item to choose.
Only used if \code{samples} is \code{TRUE}.}

\item{raw}{A \code{\link{logical}} scalar: should the raw proxy data be displayed
instead of the mean estimates?}

\item{IQR}{A \code{\link{logical}} scalar: should the IQR be displayed instead of the
95\% confidence intervals?}

\item{x, object}{A \code{\linkS4class{ProxyRecord}} object.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\value{
\itemize{
\item \code{proxy()} returns an \code{\linkS4class{ProxyRecord}} object.
\item \code{autoplot()} returns a \code{\link[ggplot2:ggplot]{ggplot}} object.
\item \code{plot()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}
}
\description{
Represents layer-counted proxy records as sequences of probability
distributions on absolute, error-free time axes.
}
\examples{
\dontrun{
## Get NGRIP records
data("ngrip2010", package = "folio")
ngrip2010 <- subset(ngrip2010, !is.na(MCE))

## Replicate fig. 3d from Boers et al. (2017)
ngrip_error <- proxy(
  depth = ngrip2010$depth,
  proxy = ngrip2010$delta,
  proxy_error = 0.01,
  time = ngrip2010$age,
  time_error = ngrip2010$MCE,
  grid = 0.01,
  calendar = "b2k"
)

plot(ngrip_error)
}
}
\references{
Boers, N., Goswami, B. & Ghil, M. (2017). A Complete Representation of
Uncertainties in Layer-Counted Paleoclimatic Archives. \emph{Climate of the
Past}, 13(9): 1169-1180. \doi{10.5194/cp-13-1169-2017}.
}
\seealso{
Other age-depth modeling tools: 
\code{\link{interpolate}()}
}
\author{
N. Frerebeau
}
\concept{age-depth modeling tools}
