% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ProxyRecord-class}
\alias{ProxyRecord-class}
\alias{.ProxyRecord}
\title{Proxy Record}
\description{
An S4 class to store proxy records.
}
\section{Slots}{

\describe{
\item{\code{year}}{A \code{\link{numeric}} vector giving the time points at which the
distribution is estimated.}

\item{\code{calendar}}{A \code{\link{character}} string specifying the chronological scale
(either "\code{CE}", "\code{BP}" or "\code{b2k}").}
}}

\section{Coerce}{

In the code snippets below, \code{x} is an \code{ProxyRecord} object.
\describe{
\item{\code{as.data.frame(x)}}{Coerces to a \code{\link{data.frame}}.}
}
}

\seealso{
Other classes: 
\code{\link{ActivityEvents-class}},
\code{\link{CumulativeEvents-class}},
\code{\link{MCMC}},
\code{\link{OccurrenceEvents-class}},
\code{\link{PhasesMCMC-class}},
\code{\link{RECE-class}},
\code{\link{RateOfChange-class}},
\code{\link{TimeRange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
