% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RateOfChange-class}
\alias{RateOfChange-class}
\alias{.RateOfChange}
\title{Rate of Change}
\description{
An S4 class to store the result of a \code{\link[=roc]{rate of change}} estimation.
}
\section{Slots}{

\describe{
\item{\code{year}}{A \code{\link{numeric}} vector giving the time points at which the
rate fo change is estimated.}

\item{\code{estimate}}{A \code{\link{numeric}} vector giving the estimation of the
rate of change.}

\item{\code{calendar}}{A \code{\link{character}} string specifying the chronological scale
(either "\code{BP}" or "\code{CE}").}

\item{\code{hash}}{A \code{\link{character}} string giving the cryptographical hash of the
original data file.}
}}

\section{Coerce}{

In the code snippets below, \code{x} is an \code{RateOfChange} object.
\describe{
\item{\code{as.data.frame(x)}}{Coerces to a \code{\link{data.frame}}.}
}
}

\seealso{
Other classes: 
\code{\link{ActivityEvents-class}},
\code{\link{CumulativeEvents-class}},
\code{\link{MCMC}},
\code{\link{OccurrenceEvents-class}},
\code{\link{PhasesMCMC-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
