% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/read.R
\docType{methods}
\name{read}
\alias{read}
\alias{read_oxcal}
\alias{read_oxcal-method}
\alias{read_bcal}
\alias{read_bcal-method}
\alias{read_chronomodel}
\alias{read_chronomodel-method}
\alias{read_oxcal,character-method}
\alias{read_bcal,character-method}
\alias{read_chronomodel,character-method}
\title{Read MCMC Output}
\usage{
read_oxcal(file, ...)

read_bcal(file, ...)

read_chronomodel(file, ...)

\S4method{read_oxcal}{character}(file, BP = FALSE)

\S4method{read_bcal}{character}(file, bin_width = 1)

\S4method{read_chronomodel}{character}(file, BP = FALSE, phases = FALSE, sep = ",", dec = ".")
}
\arguments{
\item{file}{the name of the file which the data are to be read from.
    Each row of the table appears as one line of the file.  If it does
    not contain an \emph{absolute} path, the file name is
    \emph{relative} to the current working directory,
    \code{\link{getwd}()}. Tilde-expansion is performed where supported.
    This can be a compressed file (see \code{\link{file}}).

    Alternatively, \code{file} can be a readable text-mode
    \link{connection} (which will be opened for reading if
    necessary, and if so \code{\link{close}}d (and hence destroyed) at
    the end of the function call).  (If \code{\link{stdin}()} is used,
    the prompts for lines may be somewhat confusing.  Terminate input
    with a blank line or an EOF signal, \code{Ctrl-D} on Unix and
    \code{Ctrl-Z} on Windows.  Any pushback on \code{stdin()} will be
    cleared before return.)

    \code{file} can also be a complete URL.  (For the supported URL
    schemes, see the \sQuote{URLs} section of the help for
    \code{\link{url}}.)
  }

\item{...}{Further arguments to be passed to \code{read.table}.}

\item{BP}{A \code{\link{logical}} scalar: should the data be converted from BP to
BC/AD? This should not be \code{TRUE} unless you change the default settings in
'OxCal' or 'ChronoModel'.}

\item{bin_width}{The bin width specified for the
'\href{https://bcal.shef.ac.uk/}{BCal}' calibration. Defaults to the 'BCal'
default of 1.}

\item{phases}{A \code{\link{logical}} scalar: should the data be imported as phases?}

\item{sep}{the field separator character.  Values on each line of the
    file are separated by this character.  If \code{sep = ""} (the
    default for \code{read.table}) the separator is \sQuote{white space},
    that is one or more spaces, tabs, newlines or carriage returns.}

\item{dec}{the character used in the file for decimal points.}
}
\value{
An \code{\linkS4class{MCMC}} object.
}
\description{
Read MCMC Output
}
\examples{
\dontrun{
## Import ChronoModel Output
path_zip <- system.file("extdata/chronomodel.zip", package = "ArchaeoPhases")
path_csv <- utils::unzip(path_zip, exdir = tempdir())
events <- read_chronomodel(path_csv[[1]])
phases <- read_chronomodel(path_csv[[2]])

## Import OxCal Output
path_zip <- system.file("extdata/oxcal.zip", package = "ArchaeoPhases")
path_csv <- utils::unzip(path_zip, exdir = tempdir())
oxcal <- read_chronomodel(path_csv)

is_original(oxcal, path_csv)
}
}
\references{
Bronk Ramsey, C. (2009). Bayesian Analysis of Radiocarbon Dates.
\emph{Radiocarbon}, 51(1), 337-360. \doi{10.1017/S0033822200033865}.

Buck C. E., Christen J. A. & James G. N. (1999). BCal: an on-line Bayesian
radiocarbon calibration tool. \emph{Internet Archaeology}, 7.
\doi{10.11141/ia.7.1}.

Lanos, Ph., Philippe, A. & Dufresne, Ph. (2015). Chronomodel:
Chronological Modeling of Archaeological Data using Bayesian Statistics.
URL: \url{https://www.chronomodel.fr}.
}
\seealso{
\code{\link[utils:read.table]{utils::read.table()}}

Other read: 
\code{\link{check}},
\code{\link{coda}},
\code{\link{coerce}()}
}
\author{
T. S. Dye, N. Frerebeau
}
\concept{read}
