% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RECE-class}
\alias{RECE-class}
\alias{.RECE}
\title{Radiocarbon Event Count Ensemble}
\description{
S4 classes to represent a radiocarbon event count ensemble.
}
\section{Slots}{

\describe{
\item{\code{events}}{A \code{\link{character}} vector specifying the name of the events.}

\item{\code{year}}{A \code{\link{numeric}} vector giving the time points at which the
distribution is estimated.}

\item{\code{calendar}}{A \code{\link{character}} string specifying the chronological scale
(either "\code{BP}" or "\code{CE}").}

\item{\code{hash}}{A \code{\link{character}} string giving the 32-byte MD5 hash of the
original data file.}
}}

\section{Coerce}{

In the code snippets below, \code{x} is an \code{RECE} object.
\describe{
\item{\code{as.data.frame(x)}}{Coerces to a \code{\link{data.frame}}.}
}
}

\seealso{
Other classes: 
\code{\link{ActivityEvents-class}},
\code{\link{CumulativeEvents-class}},
\code{\link{MCMC}},
\code{\link{OccurrenceEvents-class}},
\code{\link{PhasesMCMC-class}},
\code{\link{ProxyRecord-class}},
\code{\link{RateOfChange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
