% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{calendar}
\alias{BP}
\alias{b2k}
\alias{BC}
\alias{BCE}
\alias{AD}
\alias{CE}
\alias{start}
\alias{time}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{chronos}{\code{\link[chronos:gregorian]{AD}}, \code{\link[chronos:gregorian]{BC}}, \code{\link[chronos:gregorian]{BCE}}, \code{\link[chronos:gregorian]{BP}}, \code{\link[chronos:gregorian]{CE}}, \code{\link[chronos:gregorian]{b2k}}, \code{\link[chronos]{calendar}}, \code{\link[chronos]{start}}, \code{\link[chronos]{start}}, \code{\link[chronos]{time}}}
}}

