% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Classes.R
\name{reproduce}
\alias{reproduce}
\title{Reproduce an MCMC data frame}
\usage{
reproduce(x, ...)
}
\arguments{
\item{x}{An \code{archaeophases_mcmc} object.}

\item{...}{Other parameters.}
}
\description{
Reproduces a data frame from metadata held in an \code{archaeophases_mcmc}
object.
}
\author{
Thomas S. Dye, \email{tsd@tsdye.online}
}
