% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R, R/roc.R
\docType{methods}
\name{roc}
\alias{roc}
\alias{roc-method}
\alias{plot,RateOfChange,missing-method}
\alias{roc,EventsMCMC-method}
\alias{roc,CumulativeEvents-method}
\alias{roc,ActivityEvents-method}
\title{Rate of Change}
\usage{
roc(object, ...)

\S4method{plot}{RateOfChange,missing}(
  x,
  main = NULL,
  sub = NULL,
  ann = graphics::par("ann"),
  axes = TRUE,
  frame.plot = axes,
  panel.first = NULL,
  panel.last = NULL,
  ...
)

\S4method{roc}{EventsMCMC}(object, from = min(object), to = max(object), resolution = NULL)

\S4method{roc}{CumulativeEvents}(object)

\S4method{roc}{ActivityEvents}(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EventsMCMC}}, a \code{\linkS4class{CumulativeEvents}} or an
\code{\linkS4class{ActivityEvents}} object.}

\item{...}{Other \link[graphics:par]{graphical parameters} may also be passed as
arguments to this function.}

\item{x}{A \code{\linkS4class{RateOfChange}} object.}

\item{main}{A \code{\link{character}} string giving a main title for the plot.}

\item{sub}{A \code{\link{character}} string giving a subtitle for the plot.}

\item{ann}{A \code{\link{logical}} scalar: should the default annotation (title and x,
y and z axis labels) appear on the plot?}

\item{axes}{A \code{\link{logical}} scalar: should axes be drawn on the plot?}

\item{frame.plot}{A \code{\link{logical}} scalar: should a box be drawn around the
plot?}

\item{panel.first}{An an \code{expression} to be evaluated after the plot axes are
set up but before any plotting takes place. This can be useful for drawing
background grids.}

\item{panel.last}{An \code{expression} to be evaluated after plotting has taken
place but before the axes, title and box are added.}

\item{from}{A length-one \code{\link{numeric}} vector giving the earliest date to
estimate for (in years).}

\item{to}{A length-one \code{\link{numeric}} vector giving the latest date to estimate
for (in years).}

\item{resolution}{A length-one \code{\link{numeric}} vector specifying the temporal
resolution (in years) at which densities are to be estimated.
If \code{NULL} (the default), equally spaced points will be used (according to
\code{options("chronos.grid")}).}
}
\value{
\itemize{
\item \code{roc()} returns an \code{\linkS4class{RateOfChange}} object.
\item \code{plot()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}
}
\description{
Rate of Change
}
\examples{
## Coerce to MCMC
eve <- as_events(events, calendar = "CE", iteration = 1)
eve <- eve[1:10000, ]

## Tempo plot
tmp <- tempo(eve)
plot(tmp)
plot(tmp, credible = TRUE, gauss = FALSE, panel.first = grid())
plot(tmp, credible = FALSE, gauss = TRUE, panel.first = grid())

## Activity plot
act <- activity(tmp)
plot(act, panel.first = grid())
}
\seealso{
Other event tools: 
\code{\link{activity}()},
\code{\link{occurrence}()},
\code{\link{tempo}()}
}
\author{
N. Frerebeau
}
\concept{event tools}
