% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportCSV.R
\name{valid_url}
\alias{valid_url}
\title{Check if a resource can be located}
\usage{
valid_url(url_in, t = 2)
}
\arguments{
\item{url_in}{A character string.}

\item{t}{Timeout in seconds.}
}
\description{
Function retrieved from
https://stackoverflow.com/questions/52911812/check-if-url-exists-in-r
}
