% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/activity.R
\docType{methods}
\name{activity}
\alias{activity}
\alias{activity-method}
\alias{activity,EventsMCMC-method}
\alias{activity,CumulativeEvents-method}
\alias{autoplot,ActivityEvents-method}
\alias{plot,ActivityEvents,missing-method}
\alias{multiplot,ActivityEvents-method}
\title{Activity Plot}
\usage{
activity(object, ...)

\S4method{activity}{EventsMCMC}(object, from = min(object), to = max(object), resolution = NULL)

\S4method{activity}{CumulativeEvents}(object)

\S4method{autoplot}{ActivityEvents}(object, ..., fill = "grey20")

\S4method{plot}{ActivityEvents,missing}(x, fill = "grey20", ...)

\S4method{multiplot}{ActivityEvents}(...)
}
\arguments{
\item{object}{An \code{\linkS4class{EventsMCMC}} or a \code{\linkS4class{CumulativeEvents}} object.}

\item{...}{Any \code{\linkS4class{ActivityEvents}} object.}

\item{from}{A length-one \code{\link{numeric}} vector giving the earliest date to
estimate for (in years).}

\item{to}{A length-one \code{\link{numeric}} vector giving the latest date to estimate
for (in years).}

\item{resolution}{A length-one \code{\link{numeric}} vector specifying the temporal
resolution (in years) at which densities are to be estimated.
If \code{NULL} (the default), equally spaced points will be used (according to
\code{options("chronos.grid")}).}

\item{fill}{A \code{\link{character}} string specifying the colour to be used to fill
the area under the curve.}

\item{x}{An \code{\linkS4class{ActivityEvents}} object.}
}
\value{
\itemize{
\item \code{activity()} returns an \code{\linkS4class{ActivityEvents}} object.
\item \code{autoplot()} and \code{multiplot} return a \code{\link[ggplot2:ggplot]{ggplot}} object.
\item \code{plot()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}
}
\description{
Plots the first derivative of the \code{\link{tempo}} plot Bayesian estimate.
}
\examples{
## Coerce to MCMC
eve <- as_events(events, iteration = 1)
eve <- eve[1:10000, ]

## Tempo plot
tmp <- tempo(eve)
plot(tmp, credible = TRUE, gauss = FALSE)
plot(tmp, credible = FALSE, gauss = TRUE)

## Activity plot
act <- activity(tmp)
plot(act)
}
\references{
Dye, T. S. (2016). Long-term rhythms in the development of Hawaiian social
stratification. \emph{Journal of Archaeological Science}, 71: 1-9.
\doi{10.1016/j.jas.2016.05.006}.
}
\seealso{
Other event tools: 
\code{\link{occurrence}()},
\code{\link{roc}()},
\code{\link{tempo}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{event tools}
