% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/activity.R
\docType{methods}
\name{activity}
\alias{activity}
\alias{activity-method}
\alias{activity,MCMC-method}
\alias{activity,CumulativeEvents-method}
\alias{plot,ActivityEvents,missing-method}
\alias{multiplot,ActivityEvents-method}
\title{Activity Plot}
\usage{
activity(object, ...)

\S4method{activity}{MCMC}(
  object,
  time = range(object),
  n = 50 * ncol(object),
  progress = getOption("ArchaeoPhases.progress")
)

\S4method{activity}{CumulativeEvents}(object)

\S4method{plot}{ActivityEvents,missing}(x)

\S4method{multiplot}{ActivityEvents}(...)
}
\arguments{
\item{object}{An \code{\linkS4class{MCMC}} or a \code{\linkS4class{CumulativeEvents}} object.}

\item{...}{Any \code{\linkS4class{ActivityEvents}} object.}

\item{time}{A length-two \code{\link{numeric}} vector giving the earliest and latest
date to estimate for, in BC/AD years. Ignored if \code{elapsed} is \code{TRUE}.}

\item{n}{An \code{\link{integer}} specifying the number of equally spaced points at
which the density is to be estimated.}

\item{progress}{A \code{\link{logical}} scalar: should a progress bar be displayed?}

\item{x}{An \code{\linkS4class{ActivityEvents}} object.}
}
\description{
Plots the first derivative of the \code{\link{tempo}} plot Bayesian estimate.
}
\examples{
\donttest{
## Coerce to MCMC
eve <- as_mcmc(events, iteration = 1)
eve <- eve[1:10000, ]

## Tempo plot
## BP
eve_BP <- CE_to_BP(eve)
tempo_BP <- tempo(eve_BP)
plot(tempo_BP)

## CE
eve_CE <- BP_to_CE(eve_BP)
tempo_CE <- tempo(eve_CE, gauss = TRUE)
plot(tempo_CE)

## Activity plot
act <- activity(eve)
plot(act)
}
}
\references{
Dye, T. S. (2016). Long-term rhythms in the development of Hawaiian social
stratification. \emph{Journal of Archaeological Science}, 71: 1-9.
\doi{10.1016/j.jas.2016.05.006}.
}
\seealso{
Other event tools: 
\code{\link{lapse}()},
\code{\link{occurrence}()},
\code{\link{tempo}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{event tools}
