% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/activity.R
\docType{methods}
\name{activity}
\alias{activity}
\alias{activity-method}
\alias{activity,MCMC-method}
\alias{plot,ActivityEvents,missing-method}
\title{Activity Plot}
\usage{
activity(object, ...)

\S4method{activity}{MCMC}(
  object,
  elapsed = FALSE,
  origin = 1,
  time = range(object),
  n = 50 * ncol(object)
)

\S4method{plot}{ActivityEvents,missing}(x, calendar = c("BCAD", "BP"))
}
\arguments{
\item{object}{An \code{\linkS4class{MCMC}} or \code{\linkS4class{CumulativeEvents}} object.}

\item{...}{Currently not used.}

\item{elapsed}{A \code{\link{logical}} scalar: should an elapsed time scale be used?}

\item{origin}{A length-one \code{\link{numeric}} vector giving the position of the
column corresponding to the event from which elapsed time is calculated.
Only used if \code{elapsed} is \code{TRUE}.}

\item{time}{A length-two \code{\link{numeric}} vector giving the earliest and latest
date to estimate for, in BC/AD years. Ignored if \code{elapsed} is \code{TRUE}.}

\item{n}{An \code{\link{integer}} specifying the number of equally spaced points at
which the density is to be estimated.}

\item{x}{An \code{\linkS4class{ActivityEvents}} object.}

\item{calendar}{A \code{\link{character}} string specifying whether the dates
should be displayed in BP or BC/AD. It must be one of "\code{BCAD}" (the default)
or "\code{BP}". Any unambiguous substring can be given.}
}
\description{
Plots the first derivative of the \code{\link{tempo}} plot Bayesian estimate.
}
\examples{
\donttest{
## Coerce to MCMC
events <- as_mcmc(Events, iteration = 1)
events <- events[1:10000, ]

## Activity plot
act <- activity(events)
plot(act)
}
}
\references{
Dye, T. S. (2016). Long-term rhythms in the development of Hawaiian social
stratification. \emph{Journal of Archaeological Science}, 71: 1-9.
\doi{10.1016/j.jas.2016.05.006}.
}
\seealso{
Other events: 
\code{\link{occurrence}()},
\code{\link{tempo}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{events}
