% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ActivityEvents}
\alias{ActivityEvents}
\alias{ActivityEvents-class}
\alias{.ActivityEvents}
\title{Activity}
\description{
An S4 class to store the result of an \code{\link{activity}} plot.
}
\section{Slots}{

\describe{
\item{\code{calendar}}{A \code{\link{character}} string specifying the chronological scale
(either "\code{BP}" or "\code{BCAD}").}

\item{\code{hash}}{A \code{\link{character}} string giving the cryptographical hash of the
original data file.}
}}

\section{Coerce}{

In the code snippets below, \code{x} is an \code{ActivityEvents} object.
\describe{
\item{\code{as.data.frame(x)}}{Coerces to a \code{\link{data.frame}}.}
}
}

\seealso{
Other class: 
\code{\link{CumulativeEvents}},
\code{\link{MCMC}},
\code{\link{OccurrenceEvents}},
\code{\link{PhasesMCMC}}
}
\author{
N. Frerebeau
}
\concept{class}
