% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ActivityEvents-class}
\alias{ActivityEvents-class}
\alias{.ActivityEvents}
\title{Activity}
\description{
An S4 class to store the result of an \code{\link{activity}} plot.
}
\section{Slots}{

\describe{
\item{\code{year}}{A \code{\link{numeric}} vector giving the time points at which the
distribution is estimated.}

\item{\code{estimate}}{A \code{\link{numeric}} vector giving the estimation of the
distribution.}

\item{\code{calendar}}{A \code{\link{character}} string specifying the chronological scale
(either "\code{CE}", "\code{BP}" or "\code{b2k}").}

\item{\code{hash}}{A \code{\link{character}} string giving the 32-byte MD5 hash of the
original data file.}
}}

\section{Coerce}{

In the code snippets below, \code{x} is an \code{ActivityEvents} object.
\describe{
\item{\code{as.data.frame(x)}}{Coerces to a \code{\link{data.frame}}.}
}
}

\seealso{
Other classes: 
\code{\link{AgeDepthModel-class}},
\code{\link{CumulativeEvents-class}},
\code{\link{MCMC}},
\code{\link{OccurrenceEvents-class}},
\code{\link{PhasesMCMC-class}},
\code{\link{RateOfChange-class}},
\code{\link{TimeRange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
