% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenAnalyze.R
\name{allen_analyze}
\alias{allen_analyze}
\title{Analyze composite relations}
\usage{
allen_analyze(relation_1, relation_2, title, ...)
}
\arguments{
\item{relation_1}{A string denoting an Allen set.}

\item{relation_2}{A string denoting an Allen set.}

\item{title}{A string displayed as the title of the Nokel lattice.}

\item{...}{Named arguments to be passed on to \code{allen_plot()}.}
}
\value{
A layout_tbl_graph object.
}
\description{
Visualize composite Allen relations with a Nokel lattice.
}
\examples{

# Plot to the R graphics device
# allen_analyze("mDFo", "MdfO", "Composite reticulation relation")

}
\author{
Thomas S. Dye
}
