% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_composition}
\alias{allen_composition}
\title{Composition of two Allen relations}
\usage{
allen_composition(first, second)
}
\arguments{
\item{first}{the first Allen relation}

\item{second}{the second Allen relation}
}
\value{
A dataframe for input to \code{allen_plot}
}
\description{
Composition of two Allen relations
}
\author{
Thomas S. Dye
}
