% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenConcur.R
\name{allen_joint_concurrency}
\alias{allen_joint_concurrency}
\title{Joint concurrence of two or more observed intervals}
\usage{
allen_joint_concurrency(mcmc, chains, ...)
}
\arguments{
\item{mcmc}{Dataframe or archaeophases_mcmc object with the MCMC output
from a Bayesian calibration.}

\item{chains}{a list of vectors of names or indexes of columns in \code{mcmc}.}

\item{...}{Arguments to \code{multi_marginal_statistics}.}
}
\value{
foo bar
}
\description{
Estimate the age of an undated context based on the
known depositional history of associated artifacts.
}
\author{
Thomas S. Dye
}
