% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allen-mcmc.R
\name{allen_relate_intervals}
\alias{allen_relate_intervals}
\title{Relate Two or More Observed Intervals}
\usage{
allen_relate_intervals(mcmc, converse = TRUE)
}
\arguments{
\item{mcmc}{An \code{\linkS4class{EventsMCMC}} object containing the output of the
MCMC algorithm.}

\item{converse}{A \code{\link{logical}} scalar: should converse relations be observed?}
}
\value{
A \code{list} of \code{data.frame} to be passed to \code{\link[=allen_plot]{allen_plot()}}.
}
\description{
Reads MCMC output to create a data frame suitable for plotting the observed
Allen relation of two intervals.
}
\author{
T. S. Dye, N. Frerebeau
}
\keyword{internal}
