% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenIo.R
\name{allen_relation_summary}
\alias{allen_relation_summary}
\title{Summarize the relation of two phases}
\usage{
allen_relation_summary(mcmc, phases, app = "bcal", quiet = "partial")
}
\arguments{
\item{mcmc}{path to a csv file with MCMC output}

\item{phases}{a vector with four column indices representing the start
and end chains of two phases}

\item{app}{one of 'bcal', 'oxcal', or 'chronomodel' to specify which
Bayesian calibration application produced the MCMC output}

\item{quiet}{One of 'no' to allow messages and warnings,
'partial' (default) to suppress messages and allow warnings, or 'yes'
to suppress messages and warnings.}
}
\value{
a list with the following components:
}
\description{
Summarize the relation of two phases
}
\author{
Thomas S. Dye
}
