% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{allen_union}
\alias{allen_union}
\title{Union of two Allen relation sets.}
\usage{
allen_union(set_1, set_2)
}
\arguments{
\item{set_1}{The first Allen relation set or result vector}

\item{set_2}{The second Allen relation set or result vector}
}
\value{
An Allen relation set
}
\description{
Returns the union of two Allen relation sets, taking care to handle
empty sets and the sets represented by result vectors.
}
\author{
Thomas S. Dye
}
