% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/test.R
\docType{methods}
\name{apportion}
\alias{apportion}
\alias{apportion-method}
\alias{apportion,MCMC,numeric,numeric-method}
\alias{apportion,MCMC,list,missing-method}
\title{Apportioned Probabilities}
\usage{
apportion(object, from, to, ...)

\S4method{apportion}{MCMC,numeric,numeric}(object, from, to, groups = NULL)

\S4method{apportion}{MCMC,list,missing}(object, from, groups = NULL)
}
\arguments{
\item{object}{An \code{\linkS4class{MCMC}} object.}

\item{from}{A \code{\link{numeric}} vector. If \code{to} is missing, must be a \code{\link{list}} (or
a \code{\link{data.frame}}) with \code{numeric} components (columns) \code{from} and \code{to}.}

\item{to}{A \code{\link{numeric}} vector. If missing, an attempt is made to interpret
\code{from} in a suitable way.}

\item{...}{Currently not used.}

\item{groups}{A \code{\link{factor}} vector in the sense that \code{as.factor(groups)}
defines the grouping. If \code{from} is a \code{list} (or a \code{data.frame}), \code{groups}
can be a length-one vector giving the index of the grouping component
(column) of \code{from}.}
}
\value{
A \code{\link{numeric}} \code{\link{matrix}} of probabilities.
}
\description{
Apportioned Probabilities
}
\seealso{
Other tests: 
\code{\link{lapse}()},
\code{\link{older}()}
}
\author{
N. Frerebeau, A. Philippe
}
\concept{tests}
