% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchaeoPhases-package.R
\docType{package}
\name{ArchaeoPhases-package}
\alias{ArchaeoPhases-package}
\alias{_PACKAGE}
\alias{ArchaeoPhases}
\title{ArchaeoPhases: Post-processing of Markov Chain Monte Carlo Simulations for
    Chronological Modelling}
\description{
Statistical analysis of archaeological dates and groups of
    dates. This package allows to post-process Markov Chain Monte Carlo
    (MCMC) simulations from 'ChronoModel' <https://chronomodel.com/>, 'Oxcal'
    <https://c14.arch.ox.ac.uk/oxcal.html> or 'BCal'
    <https://bcal.shef.ac.uk/>. It provides functions for the estimation
    and visualization of time ranges from the posterior distribution of a
    series of dates (e.g. transition and hiatus between successive
    phases).
}
\section{Package options}{

\code{ArchaeoPhases} uses the following \code{\link[=options]{options()}} to configure behaviour:
\itemize{
\item \code{ArchaeoPhases.precision}: an \code{\link{integer}} indicating the number of decimal
places (defaults to \eqn{0}).
\item \code{ArchaeoPhases.progress}: a \code{\link{logical}} scalar specifying if progress bars
should be displayed.
}
}

\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/ArchaeoStat/ArchaeoPhases/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Thomas S. Dye \tab \emph{University of Hawai'i at Manoa, USA} \cr
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Anne Philippe \tab \emph{Université de Nantes, France} \cr
Marie-Anne Vibet \tab \emph{Université de Nantes, France} \cr
}

\strong{Package maintainer}

Anne Philippe\cr
\email{anne.philippe@univ-nantes.fr}

Laboratoire de Mathématiques Jean Leray (UMR 6629)\cr
2, rue de la Houssinière\cr
BP 92208\cr
F-44322 Nantes Cedex 3\cr
France
}
\keyword{internal}
