% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchaeoPhases-package.R
\docType{package}
\name{ArchaeoPhases-package}
\alias{ArchaeoPhases-package}
\alias{_PACKAGE}
\alias{ArchaeoPhases}
\title{ArchaeoPhases: Post-Processing of the Markov Chain Simulated by 'ChronoModel',
    'Oxcal' or 'BCal'}
\description{
Provides a list of functions for the statistical analysis of
    archaeological dates and groups of dates. It is based on the
    post-processing of the Markov Chains whose stationary distribution is
    the posterior distribution of a series of dates. Such output can be
    simulated by different applications such as 'ChronoModel'
    <https://chronomodel.com/>, 'Oxcal'
    <https://c14.arch.ox.ac.uk/oxcal.html> or 'BCal'
    <https://bcal.shef.ac.uk/>. The only requirement is to have a csv file
    containing a sample from the posterior distribution.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/ArchaeoStat/ArchaeoPhases/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Thomas S. Dye \tab \emph{University of Hawai'i at Manoa, USA} \cr
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Anne Philippe \tab \emph{Université de Nantes, France} \cr
Marie-Anne Vibet \tab \emph{Université de Nantes, France} \cr
}

\strong{Package maintainer}

Anne Philippe\cr
\email{anne.philippe@univ-nantes.fr}

Laboratoire de Mathématiques Jean Leray (UMR 6629)\cr
2, rue de la Houssinière\cr
BP 92208\cr
F-44322 Nantes Cedex 3\cr
France
}
\keyword{internal}
