% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/events.R
\docType{methods}
\name{as_events}
\alias{as_events}
\alias{as_events-method}
\alias{as_events,matrix-method}
\alias{as_events,data.frame-method}
\title{Coerce to Events}
\usage{
as_events(from, ...)

\S4method{as_events}{matrix}(from, calendar, iteration = NULL)

\S4method{as_events}{data.frame}(from, calendar, iteration = NULL)
}
\arguments{
\item{from}{from An object to be coerced.}

\item{...}{Currently not used.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the source calendar
(see \code{\link[=calendar]{calendar()}}).}

\item{iteration}{A length-one \code{\link{numeric}} vector specifying the index of the
iteration column.}
}
\value{
An \code{\linkS4class{EventsMCMC}} object.
}
\description{
Coerce to Events
}
\examples{
## Coerce to events
eve <- as_events(mcmc_events, calendar = CE(), iteration = 1)
}
\seealso{
Other event tools: 
\code{\link{activity}()},
\code{\link{elapse}()},
\code{\link{occurrence}()},
\code{\link{tempo}()}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{event tools}
