% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/bind.R
\docType{methods}
\name{bind}
\alias{bind}
\alias{cbind2-method}
\alias{cbind2,MCMC,MCMC-method}
\title{Combine two Objects}
\usage{
\S4method{cbind2}{MCMC,MCMC}(x, y)
}
\arguments{
\item{x, y}{An \code{\linkS4class{MCMC}} object.}
}
\value{
An \code{\linkS4class{MCMC}} object.
}
\description{
Combine two MCMC Objects.
}
\seealso{
Other mutators: 
\code{\link{mutator}},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
