% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/phases.R
\docType{methods}
\name{boundaries}
\alias{boundaries}
\alias{boundaries-method}
\alias{boundaries,numeric,numeric-method}
\alias{boundaries,PhasesMCMC,missing-method}
\title{Phase Time Range}
\usage{
boundaries(x, y, ...)

\S4method{boundaries}{numeric,numeric}(x, y, level = 0.95)

\S4method{boundaries}{PhasesMCMC,missing}(x, level = 0.95)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, \code{x} must be an
\code{\linkS4class{PhasesMCMC}} object.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\value{
A \code{\link{data.frame}} containing the endpoints (in years BC/AD) of the shortest
time range (at a given \code{level}).
}
\description{
Computes the shortest interval that satisfies
\eqn{P(PhaseMin < IntervalInf < IntervalSup < PhaseMax | M) = level}
for each phase.
}
\examples{
## Coerce to MCMC
events <- as_mcmc(Events, iteration = 1)
events <- events[1:10000, ]

## Compute min-max range for all chains
all <- phase(events)
head(all)

## Compute min-max range by group
phases <- phase(events, groups = list(A = c(1, 3), B = c(2, 4)))
head(phases)

## Unless otherwise specified, the phases are assumed to be unordered
get_order(phases)

## Set chronological order
## (from the older to the youngest phase)
set_order(phases) <- c("B", "A")
get_order(phases)

## Compute phase ranges
boundaries(phases)

## Compute phase transition
transition(phases)

## Compute phase hiatus
hiatus(phases)

## Compute phase duration
d <- duration(phases)
summary(d)
}
\seealso{
Other phases: 
\code{\link{duration}()},
\code{\link{hiatus}()},
\code{\link{phase}()},
\code{\link{transition}()}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{phases}
