% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/phases.R
\docType{methods}
\name{boundaries}
\alias{boundaries}
\alias{boundaries-method}
\alias{boundaries,numeric,numeric-method}
\alias{boundaries,PhasesMCMC,missing-method}
\title{Phase Time Range}
\usage{
boundaries(x, y, ...)

\S4method{boundaries}{numeric,numeric}(x, y, level = 0.95)

\S4method{boundaries}{PhasesMCMC,missing}(x, level = 0.95)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, \code{x} must be an
\code{\linkS4class{PhasesMCMC}} object.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\value{
A \code{\link{data.frame}} containing the endpoints (in years BC/AD) of the shortest
time range (at a given \code{level}).
}
\description{
Computes the shortest interval that satisfies
\eqn{P(PhaseMin < IntervalInf < IntervalSup < PhaseMax | M) = level}
for each phase.
}
\examples{
## Coerce to MCMC
eve <- as_events(events, calendar = "CE", iteration = 1)
eve <- eve[1:10000, ]

## Compute min-max range by group
## Unless otherwise specified, the phases are assumed to be unordered
pha <- phase(eve, groups = list(A = c(1, 3), B = c(2, 4)))

## Set chronological order
## (from the oldest to the youngest phase)
set_order(pha) <- c("B", "A")

## Compute phase ranges
boundaries(pha)

## Compute phase transition
transition(pha)

## Compute phase hiatus
hiatus(pha)

## Compute phase duration
dur <- duration(pha)
summary(dur)
}
\seealso{
Other phase tools: 
\code{\link{duration}()},
\code{\link{hiatus}()},
\code{\link{phase}()},
\code{\link{transition}()}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{phase tools}
