% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{calendar}
\alias{calendar}
\alias{elapse}
\alias{elapse-method}
\alias{BP_to_CE}
\alias{BP_to_CE-method}
\alias{CE_to_BP}
\alias{CE_to_BP-method}
\alias{is_BP}
\alias{is_BP-method}
\alias{is_CE}
\alias{is_CE-method}
\alias{is_BP,MCMC-method}
\alias{is_BP,PhasesMCMC-method}
\alias{is_BP,CumulativeEvents-method}
\alias{is_BP,ActivityEvents-method}
\alias{is_CE,MCMC-method}
\alias{is_CE,PhasesMCMC-method}
\alias{is_CE,CumulativeEvents-method}
\alias{is_CE,ActivityEvents-method}
\alias{elapse,MCMC-method}
\alias{BP_to_CE,numeric-method}
\alias{BP_to_CE,matrix-method}
\alias{BP_to_CE,array-method}
\alias{BP_to_CE,MCMC-method}
\alias{BP_to_CE,PhasesMCMC-method}
\alias{BP_to_CE,CumulativeEvents-method}
\alias{CE_to_BP,numeric-method}
\alias{CE_to_BP,matrix-method}
\alias{CE_to_BP,array-method}
\alias{CE_to_BP,MCMC-method}
\alias{CE_to_BP,PhasesMCMC-method}
\alias{CE_to_BP,CumulativeEvents-method}
\title{Time Scale}
\usage{
elapse(object, ...)

BP_to_CE(object)

CE_to_BP(object)

is_BP(object)

is_CE(object)

\S4method{is_BP}{MCMC}(object)

\S4method{is_BP}{PhasesMCMC}(object)

\S4method{is_BP}{CumulativeEvents}(object)

\S4method{is_BP}{ActivityEvents}(object)

\S4method{is_CE}{MCMC}(object)

\S4method{is_CE}{PhasesMCMC}(object)

\S4method{is_CE}{CumulativeEvents}(object)

\S4method{is_CE}{ActivityEvents}(object)

\S4method{elapse}{MCMC}(object, origin = 1)

\S4method{BP_to_CE}{numeric}(object)

\S4method{BP_to_CE}{matrix}(object)

\S4method{BP_to_CE}{array}(object)

\S4method{BP_to_CE}{MCMC}(object)

\S4method{BP_to_CE}{PhasesMCMC}(object)

\S4method{BP_to_CE}{CumulativeEvents}(object)

\S4method{CE_to_BP}{numeric}(object)

\S4method{CE_to_BP}{matrix}(object)

\S4method{CE_to_BP}{array}(object)

\S4method{CE_to_BP}{MCMC}(object)

\S4method{CE_to_BP}{PhasesMCMC}(object)

\S4method{CE_to_BP}{CumulativeEvents}(object)
}
\arguments{
\item{object}{An object (typically an \code{\linkS4class{MCMC}} object).}

\item{...}{Currently not used.}

\item{origin}{An \code{\link{integer}} giving the position of the column corresponding
to the event from which elapsed time is calculated.}
}
\value{
\itemize{
\item \code{elapse()}, \code{BP_to_CE()} and \code{CE_to_BP()} return an object of the same
class as \code{object}.
\item \code{is_BP()} and \code{is_CE()} return a \code{\link{logical}} scalar.
}

An object of the same sort as \code{object} with a new time scale.
}
\description{
Converts between BP and CE time scales.
}
\note{
There is no year \eqn{0} in BCE/CE scale.
}
\examples{
## Coerce to MCMC
eve <- as_mcmc(events, iteration = 1)

## Plot events
## BP
eve_BP <- CE_to_BP(eve)
plot(eve_BP)

## CE
eve_CE <- BP_to_CE(eve_BP)
plot(eve_CE)

## Elapsed origin
eve_elapse <- elapse(eve, origin = 4)
plot(eve_elapse)
}
\author{
N. Frerebeau
}
\concept{time scales}
