% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{calendar}
\alias{calendar}
\alias{BP_to_CE}
\alias{BP_to_CE-method}
\alias{BP_to_b2k}
\alias{BP_to_b2k-method}
\alias{CE_to_BP}
\alias{CE_to_BP-method}
\alias{CE_to_b2k}
\alias{CE_to_b2k-method}
\alias{b2k_to_BP}
\alias{b2k_to_BP-method}
\alias{b2k_to_CE}
\alias{b2k_to_CE-method}
\alias{is_BP}
\alias{is_BP-method}
\alias{is_CE}
\alias{is_CE-method}
\alias{is_b2k}
\alias{is_b2k-method}
\alias{is_BP,MCMC-method}
\alias{is_BP,PhasesMCMC-method}
\alias{is_BP,CumulativeEvents-method}
\alias{is_BP,ActivityEvents-method}
\alias{is_BP,OccurrenceEvents-method}
\alias{is_CE,MCMC-method}
\alias{is_CE,PhasesMCMC-method}
\alias{is_CE,CumulativeEvents-method}
\alias{is_CE,ActivityEvents-method}
\alias{is_CE,OccurrenceEvents-method}
\alias{is_b2k,MCMC-method}
\alias{is_b2k,PhasesMCMC-method}
\alias{is_b2k,CumulativeEvents-method}
\alias{is_b2k,ActivityEvents-method}
\alias{is_b2k,OccurrenceEvents-method}
\alias{BP_to_CE,numeric-method}
\alias{BP_to_CE,matrix-method}
\alias{BP_to_CE,array-method}
\alias{BP_to_CE,MCMC-method}
\alias{BP_to_CE,PhasesMCMC-method}
\alias{CE_to_BP,numeric-method}
\alias{CE_to_BP,matrix-method}
\alias{CE_to_BP,array-method}
\alias{CE_to_BP,MCMC-method}
\alias{CE_to_BP,PhasesMCMC-method}
\alias{CE_to_b2k,numeric-method}
\alias{CE_to_b2k,matrix-method}
\alias{CE_to_b2k,array-method}
\alias{CE_to_b2k,MCMC-method}
\alias{CE_to_b2k,PhasesMCMC-method}
\alias{b2k_to_BP,numeric-method}
\alias{b2k_to_BP,matrix-method}
\alias{b2k_to_BP,array-method}
\alias{b2k_to_CE,numeric-method}
\alias{b2k_to_CE,matrix-method}
\alias{b2k_to_CE,array-method}
\title{Time Scale Conversion}
\usage{
BP_to_CE(object)

BP_to_b2k(object)

CE_to_BP(object)

CE_to_b2k(object)

b2k_to_BP(object)

b2k_to_CE(object)

is_BP(object)

is_CE(object)

is_b2k(object)

\S4method{is_BP}{MCMC}(object)

\S4method{is_BP}{PhasesMCMC}(object)

\S4method{is_BP}{CumulativeEvents}(object)

\S4method{is_BP}{ActivityEvents}(object)

\S4method{is_BP}{OccurrenceEvents}(object)

\S4method{is_CE}{MCMC}(object)

\S4method{is_CE}{PhasesMCMC}(object)

\S4method{is_CE}{CumulativeEvents}(object)

\S4method{is_CE}{ActivityEvents}(object)

\S4method{is_CE}{OccurrenceEvents}(object)

\S4method{is_b2k}{MCMC}(object)

\S4method{is_b2k}{PhasesMCMC}(object)

\S4method{is_b2k}{CumulativeEvents}(object)

\S4method{is_b2k}{ActivityEvents}(object)

\S4method{is_b2k}{OccurrenceEvents}(object)

\S4method{BP_to_CE}{numeric}(object)

\S4method{BP_to_CE}{matrix}(object)

\S4method{BP_to_CE}{array}(object)

\S4method{BP_to_CE}{MCMC}(object)

\S4method{BP_to_CE}{PhasesMCMC}(object)

\S4method{CE_to_BP}{numeric}(object)

\S4method{CE_to_BP}{matrix}(object)

\S4method{CE_to_BP}{array}(object)

\S4method{CE_to_BP}{MCMC}(object)

\S4method{CE_to_BP}{PhasesMCMC}(object)

\S4method{CE_to_b2k}{numeric}(object)

\S4method{CE_to_b2k}{matrix}(object)

\S4method{CE_to_b2k}{array}(object)

\S4method{CE_to_b2k}{MCMC}(object)

\S4method{CE_to_b2k}{PhasesMCMC}(object)

\S4method{b2k_to_BP}{numeric}(object)

\S4method{b2k_to_BP}{matrix}(object)

\S4method{b2k_to_BP}{array}(object)

\S4method{b2k_to_CE}{numeric}(object)

\S4method{b2k_to_CE}{matrix}(object)

\S4method{b2k_to_CE}{array}(object)
}
\arguments{
\item{object}{An object (typically an \code{\linkS4class{MCMC}} object).}
}
\value{
\itemize{
\item \code{BP_to_CE()}, \code{BP_to_b2k()}, \code{CE_to_BP()}, \code{CE_to_b2k()}, \code{b2k_to_CE()}
and \code{b2k_to_BP()} return an object of the same sort as \code{object} with a
new time scale.
\item \code{is_BP()}, \code{is_CE()}, \code{is_b2k()} return a \code{\link{logical}} scalar.
}
}
\description{
Converts between BP (Before Present), CE (Common Era) and b2k (before 2000)
time scales.
}
\note{
There is no year \eqn{0} in BCE/CE scale.
}
\examples{
## Coerce to MCMC
eve <- as_events(events, calendar = "CE", iteration = 1)

## Plot events
## BP
eve_BP <- CE_to_BP(eve)
plot(eve_BP)

## CE
eve_CE <- BP_to_CE(eve_BP)
plot(eve_CE)

## Elapsed origin
eve_elapse <- elapse(eve, origin = 4)
plot(eve_elapse)
}
\seealso{
Other time scales: 
\code{\link{elapse}()}
}
\author{
N. Frerebeau
}
\concept{time scales}
