% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{calendar}
\alias{calendar}
\alias{elapse}
\alias{elapse-method}
\alias{BP_to_CE}
\alias{BP_to_CE-method}
\alias{CE_to_BP}
\alias{CE_to_BP-method}
\alias{b2k_to_BP}
\alias{b2k_to_BP-method}
\alias{b2k_to_CE}
\alias{b2k_to_CE-method}
\alias{is_BP}
\alias{is_BP-method}
\alias{is_CE}
\alias{is_CE-method}
\alias{is_b2k}
\alias{is_b2k-method}
\alias{is_BP,MCMC-method}
\alias{is_BP,PhasesMCMC-method}
\alias{is_BP,CumulativeEvents-method}
\alias{is_BP,ActivityEvents-method}
\alias{is_BP,OccurrenceEvents-method}
\alias{is_BP,ProxyRecord-method}
\alias{is_CE,MCMC-method}
\alias{is_CE,PhasesMCMC-method}
\alias{is_CE,CumulativeEvents-method}
\alias{is_CE,ActivityEvents-method}
\alias{is_CE,OccurrenceEvents-method}
\alias{is_CE,ProxyRecord-method}
\alias{is_b2k,MCMC-method}
\alias{is_b2k,PhasesMCMC-method}
\alias{is_b2k,CumulativeEvents-method}
\alias{is_b2k,ActivityEvents-method}
\alias{is_b2k,OccurrenceEvents-method}
\alias{is_b2k,ProxyRecord-method}
\alias{elapse,MCMC-method}
\alias{BP_to_CE,numeric-method}
\alias{BP_to_CE,matrix-method}
\alias{BP_to_CE,array-method}
\alias{BP_to_CE,MCMC-method}
\alias{BP_to_CE,PhasesMCMC-method}
\alias{BP_to_CE,CumulativeEvents-method}
\alias{CE_to_BP,numeric-method}
\alias{CE_to_BP,matrix-method}
\alias{CE_to_BP,array-method}
\alias{CE_to_BP,MCMC-method}
\alias{CE_to_BP,PhasesMCMC-method}
\alias{CE_to_BP,CumulativeEvents-method}
\alias{b2k_to_BP,numeric-method}
\alias{b2k_to_BP,matrix-method}
\alias{b2k_to_BP,array-method}
\alias{b2k_to_BP,ProxyRecord-method}
\alias{b2k_to_CE,numeric-method}
\alias{b2k_to_CE,matrix-method}
\alias{b2k_to_CE,array-method}
\alias{b2k_to_CE,ProxyRecord-method}
\title{Time Scales}
\usage{
elapse(object, ...)

BP_to_CE(object)

CE_to_BP(object)

b2k_to_BP(object)

b2k_to_CE(object)

is_BP(object)

is_CE(object)

is_b2k(object)

\S4method{is_BP}{MCMC}(object)

\S4method{is_BP}{PhasesMCMC}(object)

\S4method{is_BP}{CumulativeEvents}(object)

\S4method{is_BP}{ActivityEvents}(object)

\S4method{is_BP}{OccurrenceEvents}(object)

\S4method{is_BP}{ProxyRecord}(object)

\S4method{is_CE}{MCMC}(object)

\S4method{is_CE}{PhasesMCMC}(object)

\S4method{is_CE}{CumulativeEvents}(object)

\S4method{is_CE}{ActivityEvents}(object)

\S4method{is_CE}{OccurrenceEvents}(object)

\S4method{is_CE}{ProxyRecord}(object)

\S4method{is_b2k}{MCMC}(object)

\S4method{is_b2k}{PhasesMCMC}(object)

\S4method{is_b2k}{CumulativeEvents}(object)

\S4method{is_b2k}{ActivityEvents}(object)

\S4method{is_b2k}{OccurrenceEvents}(object)

\S4method{is_b2k}{ProxyRecord}(object)

\S4method{elapse}{MCMC}(object, origin = 1)

\S4method{BP_to_CE}{numeric}(object)

\S4method{BP_to_CE}{matrix}(object)

\S4method{BP_to_CE}{array}(object)

\S4method{BP_to_CE}{MCMC}(object)

\S4method{BP_to_CE}{PhasesMCMC}(object)

\S4method{BP_to_CE}{CumulativeEvents}(object)

\S4method{CE_to_BP}{numeric}(object)

\S4method{CE_to_BP}{matrix}(object)

\S4method{CE_to_BP}{array}(object)

\S4method{CE_to_BP}{MCMC}(object)

\S4method{CE_to_BP}{PhasesMCMC}(object)

\S4method{CE_to_BP}{CumulativeEvents}(object)

\S4method{b2k_to_BP}{numeric}(object)

\S4method{b2k_to_BP}{matrix}(object)

\S4method{b2k_to_BP}{array}(object)

\S4method{b2k_to_BP}{ProxyRecord}(object)

\S4method{b2k_to_CE}{numeric}(object)

\S4method{b2k_to_CE}{matrix}(object)

\S4method{b2k_to_CE}{array}(object)

\S4method{b2k_to_CE}{ProxyRecord}(object)
}
\arguments{
\item{object}{An object (typically an \code{\linkS4class{MCMC}} object).}

\item{...}{Currently not used.}

\item{origin}{An \code{\link{integer}} giving the position of the column corresponding
to the event from which elapsed time is calculated.}
}
\value{
\itemize{
\item \code{elapse()} returns an object of the same class as \code{object} with an elapsed
time scale.
\item \code{BP_to_CE()}, \code{CE_to_BP()}, \code{b2k_to_CE} and \code{b2k_to_BP} return an object
of the same class as \code{object}.
\item \code{is_BP()}, \code{is_CE()}, \code{is_b2k()} return a \code{\link{logical}} scalar.
}

An object of the same sort as \code{object} with a new time scale.
}
\description{
Converts between BP (Before Present), CE (Common Era) and b2k (before 2000)
time scales.
}
\note{
There is no year \eqn{0} in BCE/CE scale.
}
\examples{
## Coerce to MCMC
eve <- as_events(events, iteration = 1)

## Plot events
## BP
eve_BP <- CE_to_BP(eve)
plot(eve_BP)

## CE
eve_CE <- BP_to_CE(eve_BP)
plot(eve_CE)

## Elapsed origin
eve_elapse <- elapse(eve, origin = 4)
plot(eve_elapse)
}
\author{
N. Frerebeau
}
\concept{time scales}
