% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/read.R
\docType{methods}
\name{check}
\alias{check}
\alias{is_original}
\alias{is_original-method}
\alias{is_original,MCMC-method}
\alias{is_original,CumulativeEvents-method}
\alias{is_original,ActivityEvents-method}
\alias{is_original,OccurrenceEvents-method}
\title{Check for an Original MCMC File}
\usage{
is_original(object, ...)

\S4method{is_original}{MCMC}(object, file)

\S4method{is_original}{CumulativeEvents}(object, file)

\S4method{is_original}{ActivityEvents}(object, file)

\S4method{is_original}{OccurrenceEvents}(object, file)
}
\arguments{
\item{object}{An object (typically an \code{\linkS4class{MCMC}} object).}

\item{...}{Currently not used.}

\item{file}{Either a path to a CSV file or a connection.}
}
\value{
A \code{\link{logical}}: \code{TRUE} if the files match, \code{FALSE} otherwise.
}
\description{
Checks whether or not a file is identical to the one used to create
an object.
}
\examples{
\dontrun{
## Import ChronoModel Output
path_zip <- system.file("extdata/chronomodel.zip", package = "ArchaeoPhases")
path_csv <- utils::unzip(path_zip, exdir = tempdir())
eve <- read_chronomodel(path_csv[[1]], phases = FALSE)
pha <- read_chronomodel(path_csv[[2]], phases = TRUE)

## Import OxCal Output
path_zip <- system.file("extdata/oxcal.zip", package = "ArchaeoPhases")
path_csv <- utils::unzip(path_zip, exdir = tempdir())
oxcal <- read_chronomodel(path_csv)

is_original(oxcal, path_csv)
}
}
\seealso{
\code{\link[digest:digest]{digest::digest()}}

Other read methods: 
\code{\link{coda}},
\code{\link{coerce}()},
\code{\link{read}}
}
\author{
T. S. Dye, N. Frerebeau
}
\concept{read methods}
