% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce.R
\docType{methods}
\name{data.frame}
\alias{data.frame}
\alias{as.data.frame,CumulativeEvents-method}
\alias{as.data.frame,ActivityEvents-method}
\alias{as.data.frame,OccurrenceEvents-method}
\alias{as.data.frame,TimeRange-method}
\title{Coerce to a Data Frame}
\usage{
\S4method{as.data.frame}{CumulativeEvents}(x, ..., calendar = get_calendar())

\S4method{as.data.frame}{ActivityEvents}(x, ..., calendar = get_calendar())

\S4method{as.data.frame}{OccurrenceEvents}(x, ..., calendar = get_calendar())

\S4method{as.data.frame}{TimeRange}(x, ..., calendar = get_calendar())
}
\arguments{
\item{x}{An \R object.}

\item{...}{Further parameters to be passed to \code{\link[=data.frame]{data.frame()}}.}

\item{calendar}{A \code{\link[aion:TimeScale-class]{aion::TimeScale}} object specifying the target
calendar (see \code{\link[aion:calendar]{calendar()}}).}
}
\value{
A \code{\link{data.frame}} with an extra \code{time} column giving the (decimal) years at
which the time series was sampled.
}
\description{
Coerce to a Data Frame
}
\author{
N. Frerebeau
}
\concept{coercion methods}
