% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/phases.R
\docType{methods}
\name{duration}
\alias{duration}
\alias{duration-method}
\alias{duration,numeric,numeric-method}
\alias{duration,PhasesMCMC,missing-method}
\title{Phase Duration}
\usage{
duration(x, y, ...)

\S4method{duration}{numeric,numeric}(x, y)

\S4method{duration}{PhasesMCMC,missing}(x)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, \code{x} must be an
\code{\linkS4class{PhasesMCMC}} object.}

\item{...}{Currently not used.}
}
\description{
Phase Duration
}
\examples{
## Coerce to MCMC
events <- as_mcmc(Events, iteration = 1)
events <- events[1:10000, ]

## Compute min-max range for all chains
all <- phase(events)
all

## Compute min-max range by group
## Unless otherwise specified, the phases are assumed to be unordered
phases <- phase(events, groups = list(A = c(1, 3), B = c(2, 4)))
phases

## Set chronological order
## (from the older to the youngest phase)
set_order(phases) <- c("B", "A")
get_order(phases)

## Compute phase ranges
boundaries(phases)

## Compute phase transition
transition(phases)

## Compute phase hiatus
hiatus(phases)

## Compute phase duration
d <- duration(phases)
summary(d)
}
\seealso{
Other phases: 
\code{\link{boundaries}()},
\code{\link{hiatus}()},
\code{\link{phase}()},
\code{\link{transition}()}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{phases}
