% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/phases.R
\docType{methods}
\name{duration}
\alias{duration}
\alias{duration-method}
\alias{duration,numeric,numeric-method}
\alias{duration,PhasesMCMC,missing-method}
\title{Phase Duration}
\usage{
duration(x, y, ...)

\S4method{duration}{numeric,numeric}(x, y)

\S4method{duration}{PhasesMCMC,missing}(x)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, \code{x} must be an
\code{\linkS4class{PhasesMCMC}} object.}

\item{...}{Currently not used.}
}
\description{
Phase Duration
}
\examples{
## Coerce to MCMC
eve <- as_events(events, calendar = "CE", iteration = 1)
eve <- eve[1:10000, ]

## Compute min-max range by group
## Unless otherwise specified, the phases are assumed to be unordered
pha <- phase(eve, groups = list(A = c(1, 3), B = c(2, 4)))

## Set chronological order
## (from the oldest to the youngest phase)
set_order(pha) <- c("B", "A")

## Compute phase ranges
boundaries(pha)

## Compute phase transition
transition(pha)

## Compute phase hiatus
hiatus(pha)

## Compute phase duration
dur <- duration(pha)
summary(dur)
}
\seealso{
Other phase tools: 
\code{\link{boundaries}()},
\code{\link{phase}()},
\code{\link{transition}()}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{phase tools}
