% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DurationsMCMC-class}
\alias{DurationsMCMC-class}
\alias{.DurationsMCMC}
\title{MCMC Duration}
\description{
S4 classes to represent the output of a MCMC algorithm.
}
\note{
This class inherits from \code{\linkS4class{MCMC}}.
}
\seealso{
Other classes: 
\code{\link{ActivityEvents-class}},
\code{\link{AgeDepthModel-class}},
\code{\link{CumulativeEvents-class}},
\code{\link{EventsMCMC-class}},
\code{\link{MCMC-class}},
\code{\link{OccurrenceEvents-class}},
\code{\link{PhasesMCMC-class}},
\code{\link{TimeRange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
