% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calendar.R
\docType{methods}
\name{elapse}
\alias{elapse}
\alias{elapse-method}
\alias{elapse,MCMC-method}
\title{Elapsed Time Scale}
\usage{
elapse(object, ...)

\S4method{elapse}{MCMC}(object, origin = 1)
}
\arguments{
\item{object}{An object (typically an \code{\linkS4class{MCMC}} object).}

\item{...}{Currently not used.}

\item{origin}{An \code{\link{integer}} giving the position of the column corresponding
to the event from which elapsed time is calculated.}
}
\value{
Returns an object of the same class as \code{object} with an elapsed

An object of the same sort as \code{object} with a new time scale.
}
\description{
Elapsed Time Scale
}
\note{
There is no year \eqn{0} in BCE/CE scale.
}
\examples{
## Coerce to MCMC
eve <- as_events(events, calendar = "CE", iteration = 1)

## Plot events
## BP
eve_BP <- CE_to_BP(eve)
plot(eve_BP)

## CE
eve_CE <- BP_to_CE(eve_BP)
plot(eve_CE)

## Elapsed origin
eve_elapse <- elapse(eve, origin = 4)
plot(eve_elapse)
}
\seealso{
Other time scales: 
\code{\link{calendar}}
}
\author{
N. Frerebeau
}
\concept{time scales}
