% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{events}
\alias{events}
\title{Events}
\format{
A \code{\link{data.frame}} with 30,000 rows and 5 variables:
\describe{
\item{iter}{Iteration of the MCMC algorithm.}
\item{Event.2}{Information on event 2.}
\item{Event.1}{Information on event 1.}
\item{Event.22}{Information on event 22.}
\item{Event.12}{Information on event 12.}
}
}
\usage{
events
}
\description{
A data set containing information on the ages of four dated events.
}
\seealso{
Other datasets: 
\code{\link{phases}}
}
\concept{datasets}
\keyword{datasets}
