% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Events}
\alias{Events}
\title{Events}
\format{
A \code{\link{data.frame}} with 30,000 rows and 5 variables:
\describe{
\item{iter}{iteration of the MCMC algorithm}
\item{Event.2}{information on event 2}
\item{Event.1}{information on event 1}
\item{Event.22}{information on event 22}
\item{Event.12}{information on event 12}
}
}
\usage{
Events
}
\description{
A data set containing information on the ages of four dated events.
}
\keyword{datasets}
