% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fishpond}
\alias{Fishpond}
\title{Calibration of a fishpond chronology}
\format{
A data frame with 55,965 rows and 12 variables
\describe{
\item{Iteration}{iteration of the MCMC algorithm}
\item{beta.2..Layer.II.}{end date of Layer II}
\item{theta.5..Layer.II.}{age of dated event 5 in Layer II}
\item{theta.4..Layer.II.}{age of dated event 4 in Layer II}
\item{theta.3..Layer.II.}{age of dated event 3 in Layer II}
\item{theta.2..Layer.II.}{age of dated event 2 in Layer II}
\item{alpha.2..Layer.II.}{start date of Layer II}
\item{beta.1..Layer.III.}{end date of Layer III}
\item{theta.1..Layer.III.}{age of dated event 1 in Layer III}
\item{alpha.1..Layer.III.}{start date of Layer III}
\item{phi.1}{floating parameter}
\item{X}{superfluous column}
}
}
\usage{
Fishpond
}
\description{
A data set containing information on the ages of two fishpond deposits.
}
\keyword{datasets}
