% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/phases.R
\docType{methods}
\name{hiatus}
\alias{hiatus}
\alias{hiatus-method}
\alias{hiatus,numeric,numeric-method}
\alias{hiatus,PhasesMCMC,missing-method}
\title{Hiatus Between Successive Phases}
\usage{
hiatus(x, y, ...)

\S4method{hiatus}{numeric,numeric}(x, y, level = 0.95)

\S4method{hiatus}{PhasesMCMC,missing}(x, level = 0.95)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, \code{x} must be an
\code{\linkS4class{PhasesMCMC}} object.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\value{
A \code{\link{data.frame}} containing the endpoints (in years BC/AD) of the hiatus
between each pair of successive phases (at a given \code{level}).
}
\description{
Finds, if it exists, a hiatus between two successive phases.
}
\details{
The hiatus is the longest interval that satisfies
\eqn{P(Phase1Max < IntervalInf < IntervalSup < Phase2Min | M) = level}

This assumes that the phases are in temporal order constraint.
}
\examples{
## Coerce to MCMC
events <- as_mcmc(Events, iteration = 1)
events <- events[1:10000, ]

## Compute min-max range for all chains
all <- phase(events)
all

## Compute min-max range by group
## Unless otherwise specified, the phases are assumed to be unordered
phases <- phase(events, groups = list(A = c(1, 3), B = c(2, 4)))
phases

## Set chronological order
## (from the older to the youngest phase)
set_order(phases) <- c("B", "A")
get_order(phases)

## Compute phase ranges
boundaries(phases)

## Compute phase transition
transition(phases)

## Compute phase hiatus
hiatus(phases)

## Compute phase duration
d <- duration(phases)
summary(d)
}
\seealso{
Other phases: 
\code{\link{boundaries}()},
\code{\link{duration}()},
\code{\link{phase}()},
\code{\link{transition}()}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{phases}
