% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/phases.R, R/test.R
\docType{methods}
\name{hiatus}
\alias{hiatus}
\alias{hiatus-method}
\alias{hiatus,PhasesMCMC,missing-method}
\alias{hiatus,numeric,numeric-method}
\alias{hiatus,EventsMCMC,missing-method}
\alias{hiatus,EventsMCMC-method}
\title{Hiatus Between Two Dates}
\usage{
hiatus(x, y, ...)

\S4method{hiatus}{PhasesMCMC,missing}(x, level = 0.95)

\S4method{hiatus}{numeric,numeric}(x, y, level = 0.95)

\S4method{hiatus}{EventsMCMC,missing}(x, y)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, \code{x} must be an
\code{\linkS4class{PhasesMCMC}} or an \code{\linkS4class{EventsMCMC}} object.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\description{
Tests for the existence of a hiatus between two parameters.
}
\details{
Finds if a gap exists between two dates and returns the longest interval
that satisfies \eqn{P(x < HiatusInf < HiatusSup < y | M) = level}

The hiatus between two successive phases is the longest interval that
satisfies
\eqn{P(Phase1Max < IntervalInf < IntervalSup < Phase2Min | M) = level}
(this assumes that the phases are in temporal order constraint).
}
\section{Methods (by class)}{
\itemize{
\item \code{x = PhasesMCMC,y = missing}: Hiatus between successive phases.

\item \code{x = numeric,y = numeric}: Returns a length-three \code{\link{numeric}} vector (upper and upper
boundaries, and hiatus duration, if any).

\item \code{x = EventsMCMC,y = missing}: Hiatus between successive events.
}}

\examples{
## Coerce to MCMC
eve <- as_events(events, calendar = "CE", iteration = 1)
eve <- eve[1:10000, ]

## Test for anteriority
older(eve)

## Test for hiatus
hia <- hiatus(eve)
as.data.frame(hia)
}
\seealso{
Other tests: 
\code{\link{apportion}()},
\code{\link{older}()}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{tests}
