% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/interval.R
\docType{methods}
\name{hpdi}
\alias{hpdi}
\alias{hpdi-method}
\alias{hpdi,numeric-method}
\alias{hpdi,MCMC-method}
\title{Bayesian HPD Regions}
\usage{
hpdi(object, ...)

\S4method{hpdi}{numeric}(object, level = 0.95, CE = TRUE, ...)

\S4method{hpdi}{MCMC}(object, level = 0.95, ...)
}
\arguments{
\item{object}{A \code{\link{numeric}} vector or an \code{\linkS4class{MCMC}} object containing the
output of the MCMC algorithm for the parameter.}

\item{...}{Extra arguments to be passed to \code{\link[stats:density]{stats::density()}}.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}

\item{CE}{A \code{\link{logical}} scalar: are the data expressed in BC/AD years?}
}
\value{
A \code{\link{list}} of \code{numeric} \code{\link{matrix}} giving the lower and upper boundaries of
the HPD interval and associated probabilities.
}
\description{
Bayesian HPD Regions
}
\examples{
## Coerce to MCMC
eve <- as_events(events, calendar = "CE", iteration = 1)
eve <- eve[1:10000, ]

## CE
credible(eve, level = 0.95) # Credible interval
hpdi(eve, level = 0.95) # HPD interval

## BP
eve_BP <- CE_to_BP(eve)
credible(eve_BP, level = 0.95) # Credible interval
hpdi(eve_BP, level = 0.95) # HPD interval
}
\references{
Hyndman, R. J. (1996). Computing and graphing highest density regions.
\emph{American Statistician}, 50: 120-126. \doi{10.2307/2684423}.
}
\seealso{
\code{\link[stats:density]{stats::density()}}

Other statistics: 
\code{\link{credible}()},
\code{\link{sensitivity}()},
\code{\link{summary}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{statistics}
