% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllenDataFrame.R
\name{illustrate_allen_relations}
\alias{illustrate_allen_relations}
\title{Data for an illustrative graphic}
\usage{
illustrate_allen_relations(relations = "basic")
}
\arguments{
\item{relations}{One of:
\describe{
\item{basic}{show the 13 basic Allen relations (default);}
\item{concurrent}{show concurrent relations;}
\item{six}{show relations with distinct endpoints;}
\item{inherit}{show chronological information about a transmission
relationship from the receiver's point of view;}
\item{contribute}{show chronological information from the giver's point of view during
transmission;}
\item{stratigraphic}{show the basic stratigraphic relations established by an
observation of superposition;}
\item{anagenetic}{show the basic relations of artifact
change without branching;}
\item{cladogenetic}{show the basic relations of artifact change with branching;}
\item{contributors}{show the composite relations of two contributors to a reticulation
occurrence;}
\item{innovations}{show the composite relations of two entities produced by
innovation from a single source;}
\item{sequence}{show the composite relations of superposition
in a stratigraphic sequence; or}
\item{anagenesis}{show the composite relations of transmission during
anagenesis.}}}
}
\value{
A dataframe for input to allen_plot()
}
\description{
Create a dataframe that can be used as input for an illustrative plot.
Useful for describing the Allen operators: illustrate the full
set of Allen relations, concurrent Allen relations, and relations with
distinct endpoints (six-value set).  Also, useful for describing the
chronological domains of stratification, anagenesis, cladogenesis,
reticulation, and incorporation.
}
\references{
Harris, E. \emph{Principles of Archaeological Stratigraphy}. Second edition.
London: Academic Press.

Lyman, R. Lee and Michael J. O'Brien.  Seriation and cladistics: The
difference between anagenetic and cladogenetic evolution.  Chapter 5 in
\emph{Mapping Our Ancestors: Phylogenetic Approaches in Anthropology and
Prehistory.} New Brunswick: AldineTransaction.
}
\author{
Thomas S. Dye
}
