% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KADatesOxcal}
\alias{KADatesOxcal}
\title{Ksar Akil dates calibrated by OxCal}
\format{
A data frame with 10,000 rows and 27 variables:
\describe{
 \item{Pass}{iteration of the MCMC algorithm}
 \item{Ethelruda}{Ethelruda}
 \item{start.dated.IUP}{start dated IUP}
 \item{GrA.53000}{GrA 5300}
 \item{end.dated.IUP}{end dated IUP}
 \item{start.Ahmarian}{start Ahmarian}
 \item{GrA.57597}{GrA 57597}
 \item{GrA.53004}{GrA 53004}
 \item{GrA.57542}{GrA 57542}
 \item{GrA.54846}{GrA 54846}
 \item{GrA.57603}{GrA 57603}
 \item{GrA.57602}{GrA 57602}
 \item{GrA.53001}{GrA 53001}
 \item{Egbert}{Egbert}
 \item{GrA.54847}{GrA 54847}
 \item{GrA.57599}{GrA 57599}
 \item{GrA.57598}{GrA 57598}
 \item{GrA.57544}{GrA 57544}
 \item{end.Ahmarian}{end Ahmarian}
 \item{start.UP}{start UP}
 \item{GrA.57545}{GrA 57545}
 \item{GrA.53006}{GrA 53006}
 \item{GrA.54848}{GrA 54848}
 \item{end.UP}{end UP}
 \item{start.EPI}{start EPI}
 \item{GrA.53005}{GrA 53005}
 \item{end.EPI}{end EPI}
}
}
\usage{
KADatesOxcal
}
\description{
A data set
}
\keyword{datasets}
