% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KAPhasesChronoModel}
\alias{KAPhasesChronoModel}
\title{Ksar Akil phases calibrated by ChronoModel}
\format{
A data frame with 30,000 rows and 9 variables:
\describe{
 \item{iter}{iteration of the MCMC algorithm}
 \item{EPI.alpha}{start date of EPI}
 \item{EPI.beta}{end date of EPI}
 \item{UP.alpha}{start date of UP}
 \item{UP.beta}{end date of UP}
 \item{Ahmarian.alpha}{start date of Ahmarian}
 \item{Ahmarian.beta}{end date of Ahmarian}
 \item{IUP.alpha}{start date of IUP}
 \item{IUP.beta}{end date of IUP}
}
}
\usage{
KAPhasesChronoModel
}
\description{
A data set
}
\keyword{datasets}
