% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MCMC}
\alias{MCMC}
\alias{MCMC-class}
\alias{.MCMC}
\alias{EventsMCMC-class}
\alias{.EventsMCMC}
\title{MCMC Events}
\description{
S4 classes to represent the output of a MCMC algorithm.
}
\section{Slots}{

\describe{
\item{\code{phases}}{A \code{\link{factor}} vector specifying the name of the phases.}

\item{\code{calendar}}{A \code{\link{character}} string specifying the chronological scale
(either "\code{BP}" or "\code{BCAD}").}

\item{\code{hash}}{A \code{\link{character}} string giving the cryptographical hash of the
original data file.}
}}

\note{
These classes inherit from \code{\link{matrix}}.
}
\section{Subset}{

In the code snippets below, \code{x} is a \verb{*MCMC} object.
\describe{
\item{\code{x[[i]]}}{Extracts a single date/event (one chain) selected by
subscript \code{i}. \code{i} is a length-one \code{\link{numeric}} or \code{\link{character}} vector.}
}
}

\seealso{
Other class: 
\code{\link{ActivityEvents}},
\code{\link{CumulativeEvents}},
\code{\link{OccurrenceEvents}},
\code{\link{PhasesMCMC}}
}
\author{
N. Frerebeau
}
\concept{class}
