% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MCMC}
\alias{MCMC}
\alias{MCMC-class}
\alias{.MCMC}
\alias{EventsMCMC-class}
\alias{.EventsMCMC}
\title{MCMC Events}
\description{
S4 classes to represent the output of a MCMC algorithm.
}
\section{Slots}{

\describe{
\item{\code{events}}{A \code{\link{character}} vector specifying the name of the events.}

\item{\code{calendar}}{A \code{\link{character}} string specifying the chronological scale
(either "\code{BP}", "\code{CE}" or "\code{b2k}").}

\item{\code{hash}}{A \code{\link{character}} string giving the 32-byte MD5 hash of the
original data file.}
}}

\note{
These classes inherit from \code{\link{matrix}}.
}
\section{Subset}{

In the code snippets below, \code{x} is a \verb{*MCMC} object.
\describe{
\item{\code{x[[i]]}}{Extracts a single date/event (one chain) selected by
subscript \code{i}. \code{i} is a length-one \code{\link{numeric}} or \code{\link{character}} vector.}
}
}

\seealso{
Other classes: 
\code{\link{ActivityEvents-class}},
\code{\link{CumulativeEvents-class}},
\code{\link{OccurrenceEvents-class}},
\code{\link{PhasesMCMC-class}},
\code{\link{ProxyRecord-class}},
\code{\link{RateOfChange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
