% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutators.R
\docType{methods}
\name{names}
\alias{names}
\alias{names,MCMC-method}
\alias{names<-,MCMC-method}
\alias{names,PhasesMCMC-method}
\alias{names<-,PhasesMCMC-method}
\title{The Names of an Object}
\usage{
\S4method{names}{MCMC}(x)

\S4method{names}{MCMC}(x) <- value

\S4method{names}{PhasesMCMC}(x)

\S4method{names}{PhasesMCMC}(x) <- value
}
\arguments{
\item{x}{An object from which to get or set names.}

\item{value}{A possible value for the names of \code{x}.}
}
\value{
An object of the same sort as \code{x} with the new names assigned.
}
\description{
Get or set the names of an object.
}
\seealso{
Other mutators: 
\code{\link{bind}},
\code{\link{data.frame}},
\code{\link{sort}()},
\code{\link{sort.list}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{mutators}
