% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/occurrence.R
\docType{methods}
\name{occurrence}
\alias{occurrence}
\alias{occurrence-method}
\alias{occurrence,MCMC-method}
\alias{plot,OccurrenceEvents,missing-method}
\title{Occurrence Plot}
\usage{
occurrence(object, ...)

\S4method{occurrence}{MCMC}(object, interval = c("ci", "hpdi"), level = 0.95)

\S4method{plot}{OccurrenceEvents,missing}(x)
}
\arguments{
\item{object}{An \code{\linkS4class{MCMC}} object.}

\item{...}{Currently not used.}

\item{interval}{A \code{\link{character}} string specifying the confidence interval to
be drawn. It must be one of "\code{ci}" (credible interval; the default)
or "\code{hpdi}" (highest posterior density interval). Any unambiguous substring
can be given.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}

\item{x}{A \code{\linkS4class{OccurrenceEvents}} object.}
}
\value{
An \code{\linkS4class{OccurrenceEvents}} object.
}
\description{
A statistical graphic designed for the archaeological study of when
events of a specified kind occurred.
}
\details{
If we have \eqn{k} events, then we can estimate the calendar date \eqn{t}
corresponding to the smallest date such that the number of events observed
before \eqn{t} is equal to \eqn{k}.

The \code{occurrence()} estimates these occurrences and gives the credible
interval or the highest posterior density (HPD) region for a given \code{level}
of confidence.
}
\examples{
\donttest{
## Coerce to MCMC
eve <- as_mcmc(events, iteration = 1)
eve <- eve[1:10000, ]

## Occurrence plot
occ_ci <- occurrence(eve, interval = "ci")
plot(occ_ci)

occ_hpdi <- occurrence(eve, interval = "hpdi")
plot(occ_hpdi)
}
}
\seealso{
Other event tools: 
\code{\link{activity}()},
\code{\link{lapse}()},
\code{\link{tempo}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{event tools}
