% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/occurrence.R
\docType{methods}
\name{occurrence}
\alias{occurrence}
\alias{occurrence-method}
\alias{occurrence,EventsMCMC-method}
\alias{autoplot,OccurrenceEvents-method}
\alias{plot,OccurrenceEvents,missing-method}
\title{Occurrence Plot}
\usage{
occurrence(object, ...)

\S4method{occurrence}{EventsMCMC}(object, interval = c("ci", "hpdi"), level = 0.95)

\S4method{autoplot}{OccurrenceEvents}(object, ..., colour = "black")

\S4method{plot}{OccurrenceEvents,missing}(x, colour = "black", ...)
}
\arguments{
\item{object}{An \code{\linkS4class{EventsMCMC}} object.}

\item{...}{Currently not used.}

\item{interval}{A \code{\link{character}} string specifying the confidence interval to
be drawn. It must be one of "\code{ci}" (credible interval; the default)
or "\code{hpdi}" (highest posterior density interval). Any unambiguous substring
can be given.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}

\item{colour}{A \code{\link{character}} string specifying the colour of the segments.}

\item{x}{An \code{\linkS4class{OccurrenceEvents}} object.}
}
\value{
\itemize{
\item \code{occurrence()} returns an \code{\linkS4class{OccurrenceEvents}} object.
\item \code{autoplot()} and \code{multiplot} return a \code{\link[ggplot2:ggplot]{ggplot}} object.
\item \code{plot()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}

An \code{\linkS4class{OccurrenceEvents}} object.
}
\description{
A statistical graphic designed for the archaeological study of when
events of a specified kind occurred.
}
\details{
If we have \eqn{k} events, then we can estimate the calendar date \eqn{t}
corresponding to the smallest date such that the number of events observed
before \eqn{t} is equal to \eqn{k}.

The \code{occurrence()} estimates these occurrences and gives the credible
interval or the highest posterior density (HPD) region for a given \code{level}
of confidence.
}
\examples{
## Coerce to MCMC
eve <- as_events(events, calendar = "CE", iteration = 1)
eve <- eve[1:10000, ]

## Occurrence plot
occ_ci <- occurrence(eve, interval = "ci")
plot(occ_ci)

occ_hpdi <- occurrence(eve, interval = "hpdi")
plot(occ_hpdi)
}
\seealso{
Other event tools: 
\code{\link{activity}()},
\code{\link{rec}},
\code{\link{roc}()},
\code{\link{tempo}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{event tools}
