% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/occurrence.R
\docType{methods}
\name{occurrence}
\alias{occurrence}
\alias{occurrence-method}
\alias{occurrence,MCMC-method}
\alias{plot,OccurrenceEvents,missing-method}
\title{Occurrence Plot}
\usage{
occurrence(object, ...)

\S4method{occurrence}{MCMC}(
  object,
  interval = c("ci", "hpdi"),
  level = 0.95,
  elapsed = FALSE,
  origin = 1
)

\S4method{plot}{OccurrenceEvents,missing}(x, calendar = c("BCAD", "BP"))
}
\arguments{
\item{object}{An \code{\linkS4class{MCMC}} or \code{\linkS4class{OccurrenceEvents}} object.}

\item{...}{Currently not used.}

\item{interval}{A \code{\link{character}} string specifying the confidence interval to
be drawn. It must be one of "\code{ci}" (credible interval; the default)
or "\code{hpdi}" (highest posterior density interval). Any unambiguous substring
can be given.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}

\item{elapsed}{A \code{\link{logical}} scalar: should an elapsed time scale be used?}

\item{origin}{A length-one \code{\link{numeric}} vector giving the position of the
column corresponding to the event from which elapsed time is calculated.
Only used if \code{elapsed} is \code{TRUE}.}

\item{x}{A \code{\linkS4class{OccurrenceEvents}} object.}

\item{calendar}{A \code{\link{character}} string specifying whether the dates
should be displayed in BP or BC/AD. It must be one of "\code{BCAD}" (the default)
or "\code{BP}". Any unambiguous substring can be given.}
}
\value{
An \code{\linkS4class{OccurrenceEvents}} object.
}
\description{
A statistical graphic designed for the archaeological study of when
events of a specified kind occurred.
}
\details{
If we have \eqn{k} events, then we can estimate the calendar date \eqn{t}
corresponding to the smallest date such that the number of events observed
before \eqn{t} is equal to \eqn{k}.

The \code{occurrence()} estimates these occurrences and gives the credible
interval or the highest posterior density (HPD) region for a given \code{level}
of confidence.
}
\examples{
\donttest{
## Coerce to MCMC
events <- as_mcmc(Events, iteration = 1)
events <- events[1:10000, ]

## Occurrence plot
occ_ci <- occurrence(events, interval = "ci")
plot(occ_ci, calendar = "BCAD")

occ_hpdi <- occurrence(events, interval = "hpdi")
plot(occ_hpdi, calendar = "BP")
}
}
\seealso{
Other events: 
\code{\link{activity}()},
\code{\link{tempo}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{events}
