% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/test.R
\docType{methods}
\name{older}
\alias{older}
\alias{test_older}
\alias{test_older-method}
\alias{test_older,numeric,numeric-method}
\alias{test_older,MCMC-method}
\title{Bayesian Test for Anteriority/Posteriority Between Two Parameters}
\usage{
test_older(x, y, ...)

\S4method{test_older}{numeric,numeric}(x, y, ...)
}
\arguments{
\item{x}{A \code{\link{numeric}} vector giving the output of the MCMC algorithm for the
first parameter.}

\item{y}{A \code{\link{numeric}} vector giving the output of the MCMC algorithm for the
second parameter.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{numeric}} vector (the posterior probability of the assumption:
"event \code{x} is older than event \code{y}").
}
\description{
This function estimates the posterior probability that event \code{x} is older
than event \code{b} using the output of the MCMC algorithm. This provides a
Bayesian test for checking the following assumption: "Event \code{x} is older than
event \code{y}".
}
\details{
For a given output of MCMC algorithm, this function estimates the posterior
probability of the event \eqn{x < y} by the relative frequency of the event
"the value of event \code{x} is less than the value of event \code{y}" in the
simulated Markov chain.
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{tests}
