% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{oxcal_calibrate}
\alias{oxcal_calibrate}
\title{14C Calibration with OxCal}
\usage{
oxcal_calibrate(names, dates, errors, curve = "IntCal20")
}
\arguments{
\item{names}{A \code{\link{character}} vector specifying the names of the dates (e.g.
laboratory codes).}

\item{dates}{A \code{\link{numeric}} vector giving the BP dates to be calibrated.}

\item{errors}{A \code{\link{numeric}} vector giving the standard deviation of the dates
to be calibrated.}

\item{curve}{A \code{\link{character}} string specifying the calibration curve to be
used.}
}
\description{
14C Calibration with OxCal
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_download}()},
\code{\link{oxcal_execute}()},
\code{\link{oxcal_parse}()},
\code{\link{oxcal_setup}()}
}
\author{
N. Frerebeau
}
\concept{OxCal tools}
