% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{oxcal_calibrate}
\alias{oxcal_calibrate}
\title{14C Calibration with OxCal}
\usage{
oxcal_calibrate(names, dates, errors, curve = "IntCal20")
}
\arguments{
\item{names}{A \code{\link{character}} vector specifying the names of the dates (e.g.
laboratory codes).}

\item{dates}{A \code{\link{numeric}} vector giving the BP dates to be calibrated.}

\item{errors}{A \code{\link{numeric}} vector giving the standard deviation of the dates
to be calibrated.}

\item{curve}{A \code{\link{character}} string specifying the calibration curve to be
used.}
}
\value{
A \code{\link{data.frame}} with the following columns:
\describe{
\item{\code{name}}{}
\item{\code{type}}{}
\item{\code{date}}{}
\item{\code{error}}{}
\item{\code{range}}{}
\item{\code{likelihood}}{}
}
}
\description{
14C Calibration with OxCal
}
\examples{
\dontrun{
## Download OxCal
oxcal_setup()

## Calibrate 14C dates
cal <- oxcal_calibrate(
  names = c("X","Y"),
  dates = c(5000, 2500),
  errors = c(25, 50)
)
cal

## Custom script
scr <- 'Plot()
 {
  R_Date("X",5000,25);
  R_Date("Y",2500,50);
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)
res
}
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_download}()},
\code{\link{oxcal_execute}()},
\code{\link{oxcal_parse}()},
\code{\link{oxcal_setup}()}
}
\author{
N. Frerebeau
}
\concept{OxCal tools}
