% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{oxcal_download}
\alias{oxcal_download}
\title{Download OxCal}
\usage{
oxcal_download(path = NULL)
}
\arguments{
\item{path}{A \code{\link{character}} string specifying the directory to extract files
to. It will be created if necessary (see \code{\link[utils:unzip]{utils::unzip()}}).}
}
\value{
Invisibly returns \code{path}.
}
\description{
Download OxCal
}
\note{
Adapted from \code{\link[oxcAAR:quickSetupOxcal]{oxcAAR::quickSetupOxcal()}}.
}
\examples{
\dontrun{
## Download OxCal
oxcal_setup()

## Calibrate 14C dates
cal <- oxcal_calibrate(
  names = c("X","Y"),
  dates = c(5000, 4500),
  errors = c(45, 35)
)
plot(cal)

## Custom script
scr <- 'Plot()
 {
  Sequence("Sequence1")
  {
   Boundary("Begin");
   Phase("Phase1")
   {
    R_Date("Lab-1",5000,25);
    R_Date("Lab-2",4900,37);
   };
   Boundary("Between");
   Phase("Phase2")
   {
    R_Date("Lab-3",4800,43);
   };
   Boundary("End");
  };
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)
plot(res)
as.data.frame(res)
}
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_calibrate}()},
\code{\link{oxcal_execute}()},
\code{\link{oxcal_parse}()},
\code{\link{oxcal_setup}()}
}
\author{
N. Frerebeau
}
\concept{OxCal tools}
