% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{oxcal_parse}
\alias{oxcal_parse}
\title{Read and Parse OxCal Output}
\usage{
oxcal_parse(file)
}
\arguments{
\item{file}{A \code{\link{character}} string naming a JavaScript file which the data
are to be read from (typically returned by \code{\link[=oxcal_execute]{oxcal_execute()}}).}
}
\value{
A \code{\link{list}} with the following elements:
\describe{
\item{\code{ocd}}{A \code{list} holding the ranges, probability distributions, etc.
for each parameter.}
\item{\code{model}}{A \code{list} containing information about the model.}
\item{\code{calib}}{A \code{list} containing information about the calibration
curve.}
}
}
\description{
Read and Parse OxCal Output
}
\examples{
\dontrun{
## Download OxCal
oxcal_setup()

## Calibrate 14C dates
cal <- oxcal_calibrate(
  names = c("X","Y"),
  dates = c(5000, 4500),
  errors = c(45, 35)
)
plot(cal, curve = FALSE)
plot(cal, curve = TRUE)

## Custom script
scr <- 'Plot()
 {
  Sequence("Sequence1")
  {
   Boundary("Begin");
   Phase("Phase1")
   {
    R_Date("Lab-1",5000,25);
    R_Date("Lab-2",4900,37);
   };
   Boundary("Between");
   Phase("Phase2")
   {
    R_Date("Lab-3",4800,43);
   };
   Boundary("End");
  };
 };'
out <- oxcal_execute(scr)
res <- oxcal_parse(out)

plot(res, curve = FALSE)
plot(res, curve = TRUE)

as.data.frame(res)
}
}
\references{
\url{https://c14.arch.ox.ac.uk/oxcalhelp/hlp_analysis_file.html}
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_calibrate}()},
\code{\link{oxcal_download}()},
\code{\link{oxcal_execute}()},
\code{\link{oxcal_setup}()}
}
\author{
N. Frerebeau
}
\concept{OxCal tools}
