% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcal.R
\name{oxcal_setup}
\alias{oxcal_setup}
\title{Setup OxCal}
\usage{
oxcal_setup(path = NULL, ask = TRUE)
}
\arguments{
\item{path}{A \code{\link{character}} string specifying the directory where to find the
OxCal executable (or to extract OxCal files to).}

\item{ask}{A \code{\link{logical}} scalar: if OxCal is not installed, should the user
be asked to download it?
If \code{FALSE} and the OxCal executable cannot be found, will raise an error.
Only used if \R is being used interactively.}
}
\value{
Invisibly returns the path to the OxCal executable.
}
\description{
Setup OxCal
}
\details{
Downloads the latest version of Oxcal (if needed) and sets the executable
path correctly.
}
\note{
Adapted from \code{\link[oxcAAR:quickSetupOxcal]{oxcAAR::quickSetupOxcal()}}.
}
\seealso{
Other OxCal tools: 
\code{\link{oxcal_calibrate}()},
\code{\link{oxcal_download}()},
\code{\link{oxcal_execute}()},
\code{\link{oxcal_parse}()}
}
\author{
N. Frerebeau
}
\concept{OxCal tools}
