% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/phases.R
\docType{methods}
\name{phase}
\alias{phase}
\alias{phase-method}
\alias{phase,EventsMCMC,missing-method}
\alias{phase,EventsMCMC,list-method}
\title{Phases}
\usage{
phase(x, groups, ...)

\S4method{phase}{EventsMCMC,missing}(x)

\S4method{phase}{EventsMCMC,list}(x, groups)
}
\arguments{
\item{x}{An \code{\linkS4class{EventsMCMC}}.}

\item{groups}{A \code{\link{list}}.}

\item{...}{Currently not used.}
}
\value{
A \code{\linkS4class{PhasesMCMC}} object.
}
\description{
Constructs the minimum and maximum for a group of events (phase).
}
\note{
The default value of \code{start} or \code{end} corresponds to a CSV file exported
from \href{https://chronomodel.com/}{ChronoModel}.
}
\examples{
## Coerce to MCMC objects
## Events
(eve <- as_events(events[1:10000, ], calendar = "CE", iteration = 1))

## Phases
(pha <- as_phases(phases[1:10000, ], start = c(1, 3), calendar = "CE",
                  iteration = 1))

## Compute phases from events/dates
## Compute min-max range for all chains
pha1 <- phase(eve)

## Compute min-max range by group
## Unless otherwise specified, the phases are assumed to be unordered
pha2 <- phase(eve, groups = list(phase_1 = c(1, 3), phase_2 = c(2, 4)))

all(pha == pha2) # TRUE
}
\seealso{
Other phase tools: 
\code{\link{boundaries}()},
\code{\link{duration}()},
\code{\link{transition}()}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{phase tools}
