% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/phases.R
\docType{methods}
\name{phase}
\alias{phase}
\alias{phase-method}
\alias{get_phases}
\alias{get_phases-method}
\alias{get_order}
\alias{get_order-method}
\alias{set_order<-}
\alias{set_order-method}
\alias{is_ordered}
\alias{is_ordered-method}
\alias{phase,MCMC,missing-method}
\alias{phase,MCMC,list-method}
\alias{get_phases,MCMC-method}
\alias{set_order<-,MCMC,character-method}
\alias{set_order,MCMC,character-method}
\alias{get_order,MCMC-method}
\alias{is_ordered,MCMC-method}
\title{Phase}
\usage{
phase(x, groups, ...)

get_phases(x)

get_order(x, value)

set_order(x) <- value

is_ordered(x)

\S4method{phase}{MCMC,missing}(x)

\S4method{phase}{MCMC,list}(x, groups, ordered = FALSE)

\S4method{get_phases}{MCMC}(x)

\S4method{set_order}{MCMC,character}(x) <- value

\S4method{get_order}{MCMC}(x)

\S4method{is_ordered}{MCMC}(x)
}
\arguments{
\item{x}{An \code{\linkS4class{MCMC}} or a \code{\linkS4class{PhasesMCMC}} object.}

\item{groups}{A \code{\link{list}}.}

\item{...}{Currently not used.}

\item{value}{A possible value for the element(s) of \code{x}.}

\item{ordered}{A \code{\link{logical}} scalar: should the \code{groups} be regarded as
ordered (in the order given)?}
}
\value{
A \code{\linkS4class{PhasesMCMC}} object.
}
\description{
Constructs the minimum and maximum for a group of events (phase).
}
\examples{
## Coerce to MCMC
events <- as_mcmc(Events, iteration = 1)
events <- events[1:10000, ]

## Compute min-max range for all chains
all <- phase(events)
head(all)

## Compute min-max range by group
phases <- phase(events, groups = list(A = c(1, 3), B = c(2, 4)))
head(phases)

## Unless otherwise specified, the phases are assumed to be unordered
get_order(phases)

## Set chronological order
## (from the older to the youngest phase)
set_order(phases) <- c("B", "A")
get_order(phases)

## Compute phase ranges
boundaries(phases)

## Compute phase transition
transition(phases)

## Compute phase hiatus
hiatus(phases)

## Compute phase duration
d <- duration(phases)
summary(d)
}
\seealso{
Other phases: 
\code{\link{boundaries}()},
\code{\link{duration}()},
\code{\link{hiatus}()},
\code{\link{transition}()}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{phases}
