% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phases}
\alias{phases}
\title{Phases}
\format{
A \code{\link{data.frame}} with 30,000 rows and 5 variables:
\describe{
\item{iter}{Iteration of the MCMC algorithm.}
\item{Phase.2.alpha}{Start date of Phase 2.}
\item{Phase.2.beta}{End date of Phase 2.}
\item{Phase.1.alpha}{Start date of Phase 1.}
\item{Phase.1.beta}{End date of Phase 1.}
}
}
\usage{
phases
}
\description{
A data set containing information on the start and end dates of two phases.
}
\seealso{
Other datasets: 
\code{\link{events}}
}
\concept{datasets}
\keyword{datasets}
