% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Phases}
\alias{Phases}
\title{Phases}
\format{
A data frame with 30,000 rows and 5 variables:
\describe{
 \item{iter}{iteration of the MCMC algorithm}
 \item{Phase.2.alpha}{start date of Phase 2}
 \item{Phase.2.beta}{end date of Phase 2}
 \item{Phase.1.alpha}{start date of Phase 1}
 \item{Phase.1.beta}{end date of Phase 1}
}
}
\usage{
Phases
}
\description{
A data set containing information on the start and end dates of two phases.
}
\keyword{datasets}
