% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PhasesMCMC-class}
\alias{PhasesMCMC-class}
\alias{.PhasesMCMC}
\title{MCMC Phases}
\description{
S4 classes to represent a collection of phases.
}
\section{Slots}{

\describe{
\item{\code{phases}}{A \code{\link{character}} vector specifying the name of the phases.}

\item{\code{calendar}}{A \code{\link{character}} string specifying the chronological scale
(either "\code{CE}", "\code{BP}" or "\code{b2k}").}

\item{\code{hash}}{A \code{\link{character}} string giving the 32-byte MD5 hash of the
original data file.}
}}

\note{
Theses classes inherit from \code{\link{array}}.
}
\section{Coerce}{

In the code snippets below, \code{x} is a \code{PhasesMCMC} object.
\describe{
\item{\code{as.list(x)}}{Coerces to a \code{\link{list}}.}
}
}

\section{Subset}{

In the code snippets below, \code{x} is a \code{PhasesMCMC} object.
\describe{
\item{\code{x[[i]]}}{Extracts a single phase (two chains) selected by subscript
\code{i}. \code{i} is a length-one \code{\link{numeric}} or \code{\link{character}} vector.}
}
}

\seealso{
Other classes: 
\code{\link{ActivityEvents-class}},
\code{\link{AgeDepthModel-class}},
\code{\link{CumulativeEvents-class}},
\code{\link{DurationsMCMC-class}},
\code{\link{EventsMCMC-class}},
\code{\link{MCMC-class}},
\code{\link{OccurrenceEvents-class}},
\code{\link{RateOfChange-class}},
\code{\link{TimeRange-class}}
}
\author{
N. Frerebeau
}
\concept{classes}
\keyword{internal}
