% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot-method}
\alias{multiplot}
\alias{multiplot-method}
\alias{autoplot,MCMC-method}
\alias{plot,MCMC,missing-method}
\alias{autoplot,PhasesMCMC-method}
\alias{plot,PhasesMCMC,missing-method}
\title{Plot}
\usage{
multiplot(...)

\S4method{autoplot}{MCMC}(
  object,
  ...,
  density = TRUE,
  n = 512,
  interval = c("ci", "hpdi"),
  level = 0.95,
  decreasing = TRUE
)

\S4method{plot}{MCMC,missing}(
  x,
  density = TRUE,
  n = 512,
  interval = c("ci", "hpdi"),
  level = 0.95,
  decreasing = TRUE,
  ...
)

\S4method{autoplot}{PhasesMCMC}(
  object,
  ...,
  level = 0.95,
  n = 512,
  decreasing = TRUE,
  succession = is_ordered(object),
  facet = TRUE
)

\S4method{plot}{PhasesMCMC,missing}(
  x,
  level = 0.95,
  n = 512,
  decreasing = TRUE,
  succession = is_ordered(x),
  facet = TRUE,
  ...
)
}
\arguments{
\item{...}{Extra parameters to be passed to \code{\link[stats:density]{stats::density()}}.}

\item{object, x}{An \code{\linkS4class{MCMC}} or a \code{\linkS4class{PhasesMCMC}} object.}

\item{density}{A \code{\link{logical}} scalar: should estimated density be plotted?}

\item{n}{An \code{\link{integer}} specifying the number of equally spaced points at
which the density is to be estimated (should be a power of two). Only used
if \code{density} is \code{TRUE}.}

\item{interval}{A \code{\link{character}} string specifying the confidence interval to
be drawn. It must be one of "\code{ci}" (credible interval; the default)
or "\code{hpdi}" (highest posterior density interval). Any unambiguous substring
can be given.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}

\item{decreasing}{A \code{\link{logical}} scalar: should the sort order be decreasing?}

\item{succession}{A \code{\link{logical}} scalar: should time ranges be plotted instead
of densities?}

\item{facet}{A \code{\link{logical}} scalar: should a matrix of panels defined by phase
be drawn?}
}
\value{
\itemize{
\item \code{autoplot()} returns a \code{\link[ggplot2:ggplot]{ggplot}} object.
\item \code{plot()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}
}
\description{
Plot
}
\section{Methods (by class)}{
\itemize{
\item \code{x = MCMC,y = missing}: Plots of credible intervals or HPD regions of a series of
events.

\item \code{x = PhasesMCMC,y = missing}: Plots the characteristics of a group of events (phase).
}}

\examples{
## Coerce to MCMC
eve <- as_mcmc(events, iteration = 1)

## BP
eve_BP <- CE_to_BP(eve)
summary(eve_BP)

## CE
eve_CE <- BP_to_CE(eve_BP)
summary(eve_CE)

## Plot events
plot(eve_CE, interval = "ci", level = 0.68)
plot(eve_BP, interval = "hpdi", level = 0.68)

## Compute phases
pha <- as_phases(eve, groups = list(B = c(2, 4), A = c(1, 3)), ordered = TRUE)
summary(pha)

## Plot phases
plot(pha, succession = FALSE)
plot(pha, succession = TRUE)
}
\seealso{
\code{\link[stats:density]{stats::density()}}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{plot methods}
