% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/read.R
\docType{methods}
\name{read_bcal}
\alias{read_bcal}
\alias{read_bcal-method}
\alias{read_bcal,character-method}
\title{Read BCal Output}
\usage{
read_bcal(file, ...)

\S4method{read_bcal}{character}(file, bin_width = 1, calendar = BP())
}
\arguments{
\item{file}{the name of the file which the data are to be read from.
    Each row of the table appears as one line of the file.  If it does
    not contain an \emph{absolute} path, the file name is
    \emph{relative} to the current working directory,
    \code{\link{getwd}()}. Tilde-expansion is performed where supported.
    This can be a compressed file (see \code{\link{file}}).

    Alternatively, \code{file} can be a readable text-mode
    \link{connection} (which will be opened for reading if
    necessary, and if so \code{\link{close}}d (and hence destroyed) at
    the end of the function call).  (If \code{\link{stdin}()} is used,
    the prompts for lines may be somewhat confusing.  Terminate input
    with a blank line or an EOF signal, \code{Ctrl-D} on Unix and
    \code{Ctrl-Z} on Windows.  Any pushback on \code{stdin()} will be
    cleared before return.)

    \code{file} can also be a complete URL.  (For the supported URL
    schemes, see the \sQuote{URLs} section of the help for
    \code{\link{url}}.)
  }

\item{...}{Further arguments to be passed to \code{read.table}.}

\item{bin_width}{The bin width specified for the
\href{https://bcal.shef.ac.uk/}{BCal} calibration. Defaults to the BCal
default of 1.}

\item{calendar}{A \code{\linkS4class{TimeScale}} object specifying the calendar
(see \code{\link[chronos:calendar]{chronos::calendar()}}). It should be \code{\link[=BP]{BP()}} unless you change the
default settings in 'BCal'.}
}
\value{
An \code{\linkS4class{EventsMCMC}} object.
}
\description{
Reads MCMC output.
}
\examples{
if (requireNamespace("ArchaeoData", quietly = TRUE)) {
  ## Import BCal Output
  path_output <- system.file("bcal/fishpond.csv", package = "ArchaeoData")
  (bcal <- read_bcal(path_output))
}
}
\references{
Buck C. E., Christen J. A. & James G. N. (1999). BCal: an on-line Bayesian
radiocarbon calibration tool. \emph{Internet Archaeology}, 7.
\doi{10.11141/ia.7.1}.
}
\seealso{
\code{\link[utils:read.table]{utils::read.table()}}

Other read methods: 
\code{\link{check}},
\code{\link{read_chronomodel}},
\code{\link{read_oxcal}()}
}
\author{
T. S. Dye, N. Frerebeau
}
\concept{read methods}
