% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/read.R
\docType{methods}
\name{read_oxcal}
\alias{read_oxcal}
\alias{read_oxcal-method}
\alias{read_oxcal,character-method}
\title{Read OxCal Output}
\usage{
read_oxcal(file, ...)

\S4method{read_oxcal}{character}(file, BP = FALSE)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.
    Each row of the table appears as one line of the file.  If it does
    not contain an \emph{absolute} path, the file name is
    \emph{relative} to the current working directory,
    \code{\link{getwd}()}. Tilde-expansion is performed where supported.
    This can be a compressed file (see \code{\link{file}}).

    Alternatively, \code{file} can be a readable text-mode
    \link{connection} (which will be opened for reading if
    necessary, and if so \code{\link{close}}d (and hence destroyed) at
    the end of the function call).  (If \code{\link{stdin}()} is used,
    the prompts for lines may be somewhat confusing.  Terminate input
    with a blank line or an EOF signal, \code{Ctrl-D} on Unix and
    \code{Ctrl-Z} on Windows.  Any pushback on \code{stdin()} will be
    cleared before return.)

    \code{file} can also be a complete URL.  (For the supported URL
    schemes, see the \sQuote{URLs} section of the help for
    \code{\link{url}}.)
  }

\item{...}{Further arguments to be passed to \code{read.table}.}

\item{BP}{A \code{\link{logical}} scalar: should the data be converted from BP to
CE? This should not be \code{TRUE} unless you change the default settings in
'OxCal' or 'ChronoModel'.}
}
\value{
An \code{\linkS4class{EventsMCMC}} object.
}
\description{
Reads MCMC output.
}
\examples{
\dontrun{
## Import OxCal Output
path_output <- system.file("oxcal/ksarakil/MCMC_Sample.csv", package = "fasti")
(oxcal <- read_oxcal(path_output))
}
}
\references{
Bronk Ramsey, C. (2009). Bayesian Analysis of Radiocarbon Dates.
\emph{Radiocarbon}, 51(1), 337-360. \doi{10.1017/S0033822200033865}.
}
\seealso{
\code{\link[utils:read.table]{utils::read.table()}}

Other read methods: 
\code{\link{check}},
\code{\link{coda}},
\code{\link{coerce}()},
\code{\link{read_bcal}()},
\code{\link{read_chronomodel}}
}
\author{
T. S. Dye, N. Frerebeau
}
\concept{read methods}
