% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/rec.R
\docType{methods}
\name{rec}
\alias{rec}
\alias{rece}
\alias{rece-method}
\alias{rece,EventsMCMC-method}
\alias{autoplot,RECE-method}
\alias{plot,RECE,missing-method}
\title{Radiocarbon Event Count Ensemble}
\usage{
rece(object, ...)

\S4method{rece}{EventsMCMC}(object, resolution = 1, n = 100)

\S4method{autoplot}{RECE}(object, ...)

\S4method{plot}{RECE,missing}(x, y, ...)
}
\arguments{
\item{object}{An \code{\linkS4class{EventsMCMC}} object.}

\item{...}{Currently not used.}

\item{resolution}{A length-one \code{\link{numeric}} vector specifying the temporal
resolution. It should be a power of ten, eg. \eqn{1} (the default; annual
resolution) or \eqn{10} (decennial resolution).}

\item{n}{An \code{\link{integer}} specifying the number of item to choose randomly.}

\item{x, y}{An \code{\linkS4class{RECE}} object.}
}
\value{
\itemize{
\item \code{rece()} returns an \code{\linkS4class{RECE}} object.
\item \code{autoplot()} and \code{multiplot} return a \code{\link[ggplot2:ggplot]{ggplot}} object.
\item \code{plot()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}
}
\description{
Radiocarbon Event Count Ensemble
}
\references{
Carleton, W. C. (2021). Evaluating Bayesian Radiocarbon‐dated Event Count
(REC) Models for the Study of Long‐term Human and Environmental Processes.
\emph{Journal of Quaternary Science}, 36(1): 110‑23. \doi{10.1002/jqs.3256}.
}
\seealso{
\code{\link[=as_rece]{as_rece()}}

Other event tools: 
\code{\link{activity}()},
\code{\link{occurrence}()},
\code{\link{roc}()},
\code{\link{tempo}()}
}
\author{
N. Frerebeau
}
\concept{event tools}
