% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/roc.R
\docType{methods}
\name{roc}
\alias{roc}
\alias{roc-method}
\alias{roc,MCMC-method}
\alias{roc,CumulativeEvents-method}
\alias{roc,ActivityEvents-method}
\alias{autoplot,RateOfChange-method}
\alias{plot,RateOfChange,missing-method}
\title{Rate of Change}
\usage{
roc(object, ...)

\S4method{roc}{MCMC}(object, from = min(object), to = max(object), step = (to - from)/(250 - 1))

\S4method{roc}{CumulativeEvents}(object)

\S4method{roc}{ActivityEvents}(object)

\S4method{autoplot}{RateOfChange}(object, ..., colour = "black")

\S4method{plot}{RateOfChange,missing}(x, colour = "black", ...)
}
\arguments{
\item{object}{An \code{\linkS4class{MCMC}}, a \code{\linkS4class{CumulativeEvents}} or an
\code{\linkS4class{ActivityEvents}} object.}

\item{...}{Currently not used.}

\item{from}{A length-one \code{\link{numeric}} vector giving the earliest date to
estimate for (in years).}

\item{to}{A length-one \code{\link{numeric}} vector giving the latest date to estimate
for (in years).}

\item{step}{An \code{\link{integer}} specifying the increment at which the cumulative
distribution is to be estimated (in years).}

\item{colour}{A \code{\link{character}} string specifying the colour of the segments.}

\item{x}{A \code{\linkS4class{RateOfChange}} object.}
}
\value{
\itemize{
\item \code{roc()} returns an \code{\linkS4class{RateOfChange}} object.
\item \code{autoplot()} returns a \code{\link[ggplot2:ggplot]{ggplot}} object.
\item \code{plot()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}
}
\description{
Rate of Change
}
\examples{
## Coerce to MCMC
eve <- as_mcmc(events, iteration = 1)
eve <- eve[1:10000, ]

## Tempo plot
## BP
eve_BP <- CE_to_BP(eve)
tmp <- tempo(eve_BP)
plot(tmp)

## CE
eve_CE <- BP_to_CE(eve_BP)
tmp <- tempo(eve_CE, count = FALSE, gauss = TRUE)
plot(tmp)

## Activity plot
act <- activity(tmp)
plot(act)

## Rate of change
chg <- roc(act)
plot(chg)
}
\seealso{
Other event tools: 
\code{\link{activity}()},
\code{\link{lapse}()},
\code{\link{occurrence}()},
\code{\link{tempo}()}
}
\author{
N. Frerebeau
}
\concept{event tools}
