% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/subset.R
\docType{methods}
\name{subset}
\alias{subset}
\alias{[,MCMC-method}
\alias{[[,MCMC,ANY,missing-method}
\alias{[[,PhasesMCMC,ANY,missing-method}
\alias{[<-,MCMC-method}
\title{Extract or Replace Parts of an Object}
\usage{
\S4method{[}{MCMC}(x, i, j, ..., drop = TRUE)

\S4method{[[}{MCMC,ANY,missing}(x, i)

\S4method{[[}{PhasesMCMC,ANY,missing}(x, i)

\S4method{[}{MCMC}(x, i, j, ...) <- value
}
\arguments{
\item{x}{An object from which to extract element(s) or in which to replace
element(s).}

\item{i, j}{Indices specifying elements to extract or replace.}

\item{...}{Currently not used.}

\item{drop}{A \code{\link{logical}} scalar: should the result be coerced to
the lowest possible dimension? This only works for extracting elements,
not for the replacement.}

\item{value}{A possible value for the element(s) of \code{x}.}
}
\value{
A subsetted object.
}
\description{
Operators acting on objects to extract or replace parts.
}
\seealso{
Other mutator: 
\code{\link{mutator}}
}
\author{
N. Frerebeau
}
\concept{mutator}
