% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/summary.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,MCMC-method}
\alias{summary,PhasesMCMC-method}
\title{Marginal Summary Statistics for Multiple MCMC Chains}
\usage{
\S4method{summary}{MCMC}(object, level = 0.95)

\S4method{summary}{PhasesMCMC}(object, level = 0.95)
}
\arguments{
\item{object}{An \code{\linkS4class{MCMC}} or a \code{\linkS4class{PhasesMCMC}} object.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\value{
A \code{\link{data.frame}} where the rows correspond to the chains of interest and
columns to the following statistics:
\describe{
\item{mean}{The mean of the MCMC chain.}
\item{sd}{The standard deviation of the MCMC chain.}
\item{min}{Minimum value of the MCMC chain.}
\item{q1}{First quantile of the MCMC chain.}
\item{median}{Median of the MCMC chain.}
\item{q3}{Third quantile of the MCMC chain.}
\item{max}{Maximum value of the MCMC chain.}
\item{CI_lower}{Lower credible interval of the MCMC chain at \code{level}.}
\item{CI_upper}{Upper credible interval of the MCMC chain at \code{level}.}
}
}
\description{
Calculates summary statistics of the output of the MCMC algorithm for
multiple parameters. Results are given in calendar years (BC/AD).
}
\examples{
## Coerce to MCMC
eve <- as_mcmc(events, iteration = 1)
summary(eve)

## Plot events
plot(eve, interval = "ci", calendar = "BCAD")
plot(eve, interval = "hpdi", calendar = "BP")

## Compute phases
pha <- as_phases(eve, groups = list(B = c(2, 4), A = c(1, 3)), ordered = TRUE)
summary(pha)

## Plot phases
plot(pha, succession = FALSE, calendar = "BCAD")
plot(pha, succession = TRUE, calendar = "BCAD")
}
\seealso{
Other statistics: 
\code{\link{ci}},
\code{\link{hpdi}}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{statistics}
