% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/summary.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,MCMC-method}
\alias{summary,PhasesMCMC-method}
\title{Marginal Summary Statistics for Multiple MCMC Chains}
\usage{
\S4method{summary}{MCMC}(object, level = 0.95)

\S4method{summary}{PhasesMCMC}(object, level = 0.95)
}
\arguments{
\item{object}{An \code{\linkS4class{MCMC}} or a \code{\linkS4class{PhasesMCMC}} object.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\value{
A \code{\link{data.frame}} where the rows correspond to the chains of interest and
columns to the following statistics:
\describe{
\item{mean}{The mean of the MCMC chain.}
\item{sd}{The standard deviation of the MCMC chain.}
\item{min}{Minimum value of the MCMC chain.}
\item{q1}{First quantile of the MCMC chain.}
\item{median}{Median of the MCMC chain.}
\item{q3}{Third quantile of the MCMC chain.}
\item{max}{Maximum value of the MCMC chain.}
\item{CI_lower}{Lower boundary of the credible interval of the MCMC chain
at \code{level}.}
\item{CI_upper}{Upper boundary of the credible interval of the MCMC chain
at \code{level}.}
}
}
\description{
Calculates summary statistics of the output of the MCMC algorithm for
multiple parameters. Results are given in calendar years (BC/AD).
}
\examples{
## Coerce to MCMC
eve <- as_mcmc(events, iteration = 1)

## BP
eve_BP <- CE_to_BP(eve)
summary(eve_BP)

## CE
eve_CE <- BP_to_CE(eve_BP)
summary(eve_CE)

## Plot events
plot(eve_CE, interval = "credible", level = 0.68)
plot(eve_BP, interval = "hpdi", level = 0.68)

## Compute phases
pha <- phase(eve, groups = list(B = c(2, 4), A = c(1, 3)), ordered = TRUE)
summary(pha)

## Plot phases
plot(pha, succession = FALSE)
plot(pha, succession = TRUE)
}
\seealso{
Other statistics: 
\code{\link{credible}()},
\code{\link{hpdi}()},
\code{\link{sensitivity}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{statistics}
