% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tempo.R
\docType{methods}
\name{tempo}
\alias{tempo}
\alias{tempo-method}
\alias{tempo,MCMC-method}
\alias{autoplot,CumulativeEvents-method}
\alias{plot,CumulativeEvents,missing-method}
\alias{multiplot,CumulativeEvents-method}
\title{Tempo Plot}
\usage{
tempo(object, ...)

\S4method{tempo}{MCMC}(
  object,
  level = 0.95,
  count = TRUE,
  gauss = FALSE,
  from = min(object),
  to = max(object),
  step = (to - from)/(250 - 1)
)

\S4method{autoplot}{CumulativeEvents}(object, ...)

\S4method{plot}{CumulativeEvents,missing}(x)

\S4method{multiplot}{CumulativeEvents}(...)
}
\arguments{
\item{object}{An \code{\linkS4class{MCMC}} object.}

\item{...}{Any \code{\linkS4class{CumulativeEvents}} object.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}

\item{count}{A \code{\link{logical}} scalar: should the counting process be a number
(the default) or a probability?}

\item{gauss}{A \code{\link{logical}} scalar: should the Gaussian approximation of the
credible interval be used?}

\item{from}{A length-one \code{\link{numeric}} vector giving the earliest date to
estimate for (in years).}

\item{to}{A length-one \code{\link{numeric}} vector giving the latest date to estimate
for (in years).}

\item{step}{An \code{\link{integer}} specifying the increment at which the cumulative
distribution is to be estimated (in years).}

\item{x}{A \code{\linkS4class{CumulativeEvents}} object.}
}
\value{
\itemize{
\item \code{tempo()} returns an \code{\linkS4class{CumulativeEvents}} object.
\item \code{autoplot()} and \code{multiplot} return a \code{\link[ggplot2:ggplot]{ggplot}} object.
\item \code{plot()} is called it for its side-effects: it results in a graphic being
displayed (invisibly returns \code{x}).
}
}
\description{
A statistical graphic designed for the archaeological study of rhythms of the
long term that embodies a theory of archaeological evidence for the
occurrence of events.
}
\details{
The tempo plot is one way to measure change over time: it estimates the
cumulative occurrence of archaeological events in a Bayesian calibration.
The tempo plot yields a graphic where the slope of the plot directly
reflects the pace of change: a period of rapid change yields a steep slope
and a period of slow change yields a gentle slope. When there is no change,
the plot is horizontal. When change is instantaneous, the plot is vertical.
}
\examples{
## Coerce to MCMC
eve <- as_mcmc(events, iteration = 1)
eve <- eve[1:10000, ]

## Tempo plot
## BP
eve_BP <- CE_to_BP(eve)
tmp <- tempo(eve_BP)
plot(tmp)

## CE
eve_CE <- BP_to_CE(eve_BP)
tmp <- tempo(eve_CE, count = FALSE, gauss = TRUE)
plot(tmp)

## Activity plot
act <- activity(tmp)
plot(act)

## Rate of change
chg <- roc(act)
plot(chg)
}
\references{
Dye, T. S. (2016). Long-term rhythms in the development of Hawaiian social
stratification. \emph{Journal of Archaeological Science}, 71: 1-9.
\doi{10.1016/j.jas.2016.05.006}.
}
\seealso{
Other event tools: 
\code{\link{activity}()},
\code{\link{lapse}()},
\code{\link{occurrence}()},
\code{\link{roc}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{event tools}
