% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tempo.R
\docType{methods}
\name{tempo}
\alias{tempo}
\alias{tempo-method}
\alias{tempo,MCMC-method}
\alias{plot,CumulativeEvents,missing-method}
\title{Tempo Plot}
\usage{
tempo(object, ...)

\S4method{tempo}{MCMC}(
  object,
  level = 0.95,
  count = TRUE,
  gauss = FALSE,
  elapsed = FALSE,
  origin = 1,
  time = range(object),
  n = 50 * ncol(object)
)

\S4method{plot}{CumulativeEvents,missing}(x, calendar = c("BCAD", "BP"))
}
\arguments{
\item{object}{An \code{\linkS4class{MCMC}} or \code{\linkS4class{CumulativeEvents}} object.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}

\item{count}{A \code{\link{logical}} scalar: should the counting process be a number
(the default) or a probability?}

\item{gauss}{A \code{\link{logical}} scalar: should the Gaussian approximation of the
credible interval be used?}

\item{elapsed}{A \code{\link{logical}} scalar: should an elapsed time scale be used?}

\item{origin}{A length-one \code{\link{numeric}} vector giving the position of the
column corresponding to the event from which elapsed time is calculated.
Only used if \code{elapsed} is \code{TRUE}.}

\item{time}{A length-two \code{\link{numeric}} vector giving the earliest and latest
date to estimate for, in BC/AD years. Ignored if \code{elapsed} is \code{TRUE}.}

\item{n}{An \code{\link{integer}} specifying the number of equally spaced points at
which the cumulative distribution is to be estimated.}

\item{x}{A \code{\linkS4class{CumulativeEvents}} object.}

\item{calendar}{A \code{\link{character}} string specifying whether the dates
should be displayed in BP or BC/AD. It must be one of "\code{BCAD}" (the default)
or "\code{BP}". Any unambiguous substring can be given.}
}
\description{
A statistical graphic designed for the archaeological study of rhythms of the
long term that embodies a theory of archaeological evidence for the
occurrence of events
}
\details{
The tempo plot is one way to measure change over time: it estimates the
cumulative occurrence of archaeological events in a Bayesian calibration.
The tempo plot yields a graphic where the slope of the plot directly
reflects the pace of change: a period of rapid change yields a steep slope
and a period of slow change yields a gentle slope. When there is no change,
the plot is horizontal. When change is instantaneous, the plot is vertical.
}
\examples{
\donttest{
## Coerce to MCMC
events <- as_mcmc(Events, iteration = 1)
events <- events[1:10000, ]

## Tempo plot
tempo1 <- tempo(events)
plot(tempo1)

tempo2 <- tempo(events, gauss = TRUE)
plot(tempo2, calendar = "BP")
}
}
\references{
Dye, T. S. (2016). Long-term rhythms in the development of Hawaiian social
stratification. \emph{Journal of Archaeological Science}, 71: 1-9.
\doi{10.1016/j.jas.2016.05.006}.
}
\seealso{
Other events: 
\code{\link{activity}()},
\code{\link{occurrence}()}
}
\author{
A. Philippe, M.-A. Vibet, T. S. Dye, N. Frerebeau
}
\concept{events}
