% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/phases.R
\docType{methods}
\name{transition}
\alias{transition}
\alias{transition-method}
\alias{transition,numeric,numeric-method}
\alias{transition,PhasesMCMC,missing-method}
\title{Transition Range Between Successive Phases}
\usage{
transition(x, y, ...)

\S4method{transition}{numeric,numeric}(x, y, level = 0.95)

\S4method{transition}{PhasesMCMC,missing}(x, level = 0.95)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, \code{x} must be an
\code{\linkS4class{PhasesMCMC}} object.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\value{
A \code{\link{data.frame}} containing the endpoints (in years BC/AD) of the
transition interval for each pair of successive phases (at a given \code{level}).
}
\description{
Finds, if it exists, the shortest interval that satisfies
\eqn{P(TransitionInf < Phase1Max < Phase2Min < TransitionSup | M) = level}.
This assumes that phases are in temporal order constraint.
}
\examples{
## Coerce to MCMC
events <- as_mcmc(Events, iteration = 1)
events <- events[1:10000, ]

## Compute min-max range for all chains
all <- phase(events)
head(all)

## Compute min-max range by group
phases <- phase(events, groups = list(A = c(1, 3), B = c(2, 4)))
head(phases)

## Unless otherwise specified, the phases are assumed to be unordered
get_order(phases)

## Set chronological order
## (from the older to the youngest phase)
set_order(phases) <- c("B", "A")
get_order(phases)

## Compute phase ranges
boundaries(phases)

## Compute phase transition
transition(phases)

## Compute phase hiatus
hiatus(phases)
}
\seealso{
Other phases: 
\code{\link{boundaries}()},
\code{\link{hiatus}()},
\code{\link{phase}()}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{phases}
