% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/phases.R
\docType{methods}
\name{transition}
\alias{transition}
\alias{transition-method}
\alias{transition,numeric,numeric-method}
\alias{transition,PhasesMCMC,missing-method}
\title{Transition Range Between Successive Phases}
\usage{
transition(x, y, ...)

\S4method{transition}{numeric,numeric}(x, y, level = 0.95)

\S4method{transition}{PhasesMCMC,missing}(x, level = 0.95)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, \code{x} must be an
\code{\linkS4class{PhasesMCMC}} object.}

\item{...}{Currently not used.}

\item{level}{A length-one \code{\link{numeric}} vector giving the confidence level.}
}
\value{
A \code{\link{data.frame}} containing the endpoints (in years BC/AD) of the
transition interval for each pair of successive phases (at a given \code{level}).
}
\description{
Estimates the transition endpoints between two phases.
}
\details{
The transition is the shortest interval that satisfies
\eqn{P(IntervalInf < Phase1Max < Phase2Min < IntervalSup | M) = level}.

This assumes that the phases are in temporal order constraint.
}
\examples{
## Coerce to MCMC
eve <- as_mcmc(events, iteration = 1)
eve <- eve[1:10000, ]

## Compute min-max range by group
## Unless otherwise specified, the phases are assumed to be unordered
pha <- as_phases(eve, groups = list(A = c(1, 3), B = c(2, 4)))

## Set chronological order
## (from the oldest to the youngest phase)
set_order(pha) <- c("B", "A")

## Compute phase ranges
boundaries(pha)

## Compute phase transition
transition(pha)

## Compute phase hiatus
hiatus(pha)

## Compute phase duration
dur <- duration(pha)
summary(dur)
}
\seealso{
Other phase tools: 
\code{\link{boundaries}()},
\code{\link{duration}()},
\code{\link{hiatus}()},
\code{\link{phase}}
}
\author{
A. Philippe, M.-A. Vibet, N. Frerebeau
}
\concept{phase tools}
